% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.metf.R
\name{alpha.metf}
\alias{alpha.metf}
\title{Calculate the diversity for each sample or group}
\usage{
alpha.metf(
  otu = NULL,
  tax = NULL,
  map = NULL,
  ps = NULL,
  group = "Group",
  sampling = TRUE,
  Plot = TRUE
)
}
\arguments{
\item{otu}{Metagenome functional composition table;}

\item{tax}{Metagenomic Functional taxonomy table;}

\item{map}{A data frame or matrix containing metadata for samples. Must include at least `ID` and grouping column(s).}

\item{ps}{A `phyloseq` object. It must contain a metagenome functional composition table (`otu_table`), and sample metadata (`sample_data`).}

\item{group}{Column name for group, such as "Group".}

\item{sampling}{Sampling metagenome functional composition table with the minisize sequence count;}
}
\value{
A data frame containing alpha diversity indices for each sample, along with the sample metadata. The output includes the following indices:
\itemize{
  \item `Shannon`: Shannon diversity index.
  \item `Inv_Simpson`: Inverse Simpson diversity index.
  \item `Pielou_evenness`: Pielou's evenness index.
  \item `Simpson_evenness`: Simpson's evenness index.
  \item `Richness`: Observed richness (number of genes).
  \item `Chao1`: Chao1 richness estimator.
  \item `ACE`: ACE richness estimator.
}
}
\description{
The alpha.metf function calculates alpha diversity indices for metagenome functional composition.
This includes measures such as Shannon diversity, Pielou's evenness, and richness estimates (e.g., Chao1, ACE).
}
\examples{
\dontrun{
ps =ps.kegg \%>\% filter_OTU_ps(Top = 1000)
tab = alpha.metf(ps = ps,group = "Group")
}
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
