% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.metm.R
\name{alpha.metm}
\alias{alpha.metm}
\title{Calculate the diversity for each sample or group}
\usage{
alpha.metm(
  otu = NULL,
  tax = NULL,
  map = NULL,
  ps = NULL,
  group = "Group",
  sampling = TRUE
)
}
\arguments{
\item{otu}{OTU/ASV table;}

\item{tax}{taxonomy table;}

\item{map}{A data frame or matrix containing metadata for samples. Must include at least `ID` and grouping column(s).}

\item{ps}{A `phyloseq` object. It must contain an OTU/ASV table (`otu_table`), and sample metadata (`sample_data`).}

\item{group}{column name for group, such as "Group".}

\item{sampling}{sampling OTU/ASV table with the minisize sequence count;}
}
\value{
A data frame containing alpha diversity indices for each sample, along with the sample metadata. The output includes the following indices:
\itemize{
  \item `Shannon`: Shannon diversity index.
  \item `Inv_Simpson`: Inverse Simpson diversity index.
  \item `Pielou_evenness`: Pielou's evenness index.
  \item `Simpson_evenness`: Simpson's evenness index.
  \item `Richness`: Observed richness (number of species).
  \item `Chao1`: Chao1 richness estimator.
  \item `ACE`: ACE richness estimator.
}
}
\description{
The alpha.metm function calculates alpha diversity indices for metagenome microbial communities.
This includes measures such as Shannon diversity, Pielou's evenness, and richness estimates (e.g., Chao1, ACE).
}
\examples{
\dontrun{
tab = alpha.metm(ps = ps.16s,group = "Group")
}
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
