% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.micro.R
\name{alpha.micro}
\alias{alpha.micro}
\title{Calculate the diversity for each sample or group}
\usage{
alpha.micro(
  otu = NULL,
  tax = NULL,
  map = NULL,
  ps = NULL,
  group = "Group",
  sampling = TRUE
)
}
\arguments{
\item{otu}{OTU/ASV table;}

\item{tax}{taxonomy table;}

\item{map}{A data frame or matrix containing metadata for samples. Must include at least `ID` and grouping column(s).}

\item{ps}{A `phyloseq` object. It must contain an OTU/ASV table (`otu_table`), and sample metadata (`sample_data`).}

\item{group}{column name for group, such as "Group".}

\item{sampling}{sampling OTU/ASV table with the minisize sequence count;}
}
\value{
data frame of sample diversity indices.
}
\description{
The alpha.micro function calculates alpha diversity indices for microbial communities.
This includes measures such as Shannon diversity, Pielou's evenness, and richness estimates (e.g., Chao1, ACE).
}
\examples{
\dontrun{
tab = alpha.micro(ps = ps.16s,group = "Group")
}
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
