% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.pd.R
\name{alpha.pd.micro}
\alias{alpha.pd.micro}
\title{Calculate Phylogenetic Diversity (PD) for Samples}
\usage{
alpha.pd.micro(ps = NULL, group = "Group")
}
\arguments{
\item{ps}{A `phyloseq` object. It must contain an OTU/ASV table (`otu_table`), a phylogenetic tree (`phy_tree`), and sample metadata (`sample_data`).}

\item{group}{column name for group, such as "Group".}
}
\value{
data frame of sample pd indices.
}
\description{
This function calculates the phylogenetic diversity (PD) for each sample in a given phyloseq object.
}
\examples{
\dontrun{
library(ape)
library(picante)
tab2 = alpha.pd.micro(ps.16s)
head(tab2)
}
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
