% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.rare.line.R
\name{alpha.rare.line.micro}
\alias{alpha.rare.line.micro}
\title{Plotting rarefaction curve for each sample or group}
\usage{
alpha.rare.line.micro(
  otu = NULL,
  tax = NULL,
  map = NULL,
  ps = NULL,
  method = "Richness",
  group = "Group",
  start = 100,
  step = 100
)
}
\arguments{
\item{method}{method for calculate alpha diversity,including "observed", "chao1", "diversity_shannon", "diversity_inverse_simpson", "diversity_gini_simpson", "diversity_fisher",  "diversity_coverage", "evenness_camargo", "evenness_pielou", "evenness_simpson", "evenness_evar", "evenness_bulla", "dominance_dbp",  "dominance_dmn", "dominance_absolute","dominance_relative", "dominance_simpson", "dominance_core_abundance" ,  "dominance_gini", "rarity_log_modulo_skewness", "rarity_low_abundance", "rarity_noncore_abundance",  "rarity_rare_abundance"}

\item{start}{sampling OTU/ASV table with the start number sequence count;}

\item{step}{number of intervals for sampling}

\item{otutab}{OTU/ASV table;}

\item{metadata}{matrix or data frame, including sampleID and groupID;}

\item{groupID}{column name for groupID, such as "Group".}
}
\value{
ggplot2 object.
}
\description{
Input otutab and metadata, and manual set metadata column names.
ggplot2 show line plot, and/or standard error
}
\examples{
rare <- mean(phyloseq::sample_sums(ps.16s))/10
result = alpha.rare.line(ps = ps.16s, group = "Group", method = "Richness", start = 100, step = rare)
p2_1 <- result[[1]]
p2_1
# Plot the rarefaction curve for a single sample
p2_1 <- result[[1]]
p2_1
# Provide a data table for convenient output
raretab <- result[[2]]
head(raretab)
# Display rarefaction curves grouped by categories
p2_2 <- result[[3]]
p2_2
# Plot rarefaction curves with standard deviations by groups
p2_3 <- result[[4]]
p2_3
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
