% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann.HMDB.R
\name{ann.HMDB}
\alias{ann.HMDB}
\title{Retrieve metabolite information from the HMDB database}
\usage{
ann.HMDB(id, verbose = FALSE)
}
\arguments{
\item{id}{The metabolite ID for which information is to be retrieved.}
}
\value{
A data frame containing the metabolite information including HMDB ID, KEGG ID,
Super_class, Class, Sub_class and related details from the database.
}
\description{
The ann.HMDB function retrieves relevant information about a metabolite based
 on its ID from the HMDB (Human Metabolome Database) stored in the db.metabolites database.
}
\examples{
library(dplyr)
tax= ps.ms \%>\% vegan_tax() \%>\%as.data.frame()
id = tax$Metabolite
tax1 = ann.HMDB (id = id)
head(tax1)
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njau.edu.cn}
}
