% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann.kegg.R
\name{ann.kegg}
\alias{ann.kegg}
\title{Retrieve KEGG metabolite information based on input metabolite IDs}
\usage{
ann.kegg(id)
}
\arguments{
\item{id}{The metabolite ID for which KEGG information is to be retrieved.}
}
\value{
A data frame with KEGG IDs and matched metabolite names for input metabolite IDs.
}
\description{
This function matches the input metabolite ID to allMetabolites in the KEGG database by formatting it,
and finally outputs its storage ID in the KEGG database.
}
\examples{
library(dplyr)
tax= ps.ms \%>\% vegan_tax() \%>\%as.data.frame()
id = tax$Metabolite
tax2 = ann.kegg(id)
head(tax2)
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
