% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann.kegg2.R
\name{ann.kegg2}
\alias{ann.kegg2}
\title{Another matching mode for retrieving KEGG metabolite information based on input metabolite IDs}
\usage{
ann.kegg2(
  id,
  cache_dir = "./kegg_cache",
  max_age_days = 90,
  force_update = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{id}{The metabolite ID for which KEGG information is to be retrieved.}
}
\value{
A data frame with KEGG IDs and matched metabolite names for input metabolite IDs.
}
\description{
This function matches the input metabolite ID to common names in the KEGG database by formatting it,
and finally outputs its storage ID in the KEGG database.
}
\examples{
library(dplyr)
tax= ps.ms \%>\% vegan_tax() \%>\%as.data.frame()
id = tax$Metabolite
tax2 = ann.kegg2(id)
head(tax2)
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
