% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_rmt_scale_free2.R
\name{apply_rmt_analysis}
\alias{apply_rmt_analysis}
\title{Apply RMT analysis to find optimal threshold}
\usage{
apply_rmt_analysis(cor_matrix, plot_results = TRUE)
}
\description{
Apply RMT analysis to find optimal threshold
}
