% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buplotall.ms.R
\name{buplotall.ms}
\alias{buplotall.ms}
\title{Bubble Plot for Metabolic Pathway Enrichment Analysis}
\usage{
buplotall.ms(ps, dif.method = "wilcox")
}
\arguments{
\item{ps}{A `phyloseq` object containing metabolomics data, including the compound abundance table and sample metadata.}

\item{dif.method}{A character string specifying the statistical method for differential testing. Default is `"wilcox"` (Wilcoxon rank-sum test).}
}
\value{
A list containing:
\itemize{
  \item `plots`: A list of ggplot objects, where each plot corresponds to a pairwise comparison of groups and visualizes enriched metabolic pathways.
  \item `plotdata`: A list of data frames, where each data frame contains detailed results of the pathway enrichment analysis for a specific group comparison.
}
}
\description{
The `buplotall.ms` function performs pairwise comparisons of groups in a metabolomics dataset and generates bubble plots to visualize enriched metabolic pathways. It integrates statistical testing and pathway enrichment analysis, allowing users to explore significant pathways across groups.
}
\details{
The function performs the following steps:
\enumerate{
  \item Conducts pairwise group comparisons using the specified statistical method (`dif.method`).
  \item Identifies significantly altered compounds based on adjusted p-values.
  \item Maps significant compounds to metabolic pathways using a preloaded pathway database (`db.compound.pathway.bins`).
  \item Conducts pathway enrichment analysis using the `clusterProfiler::enricher` function.
  \item Generates bubble plots to visualize enriched pathways, where the x-axis represents the pathway coverage, the y-axis represents pathway enrichment significance (`-log10(p-value)`), and bubble size corresponds to the count of significant compounds in each pathway.
}
}
\examples{
\dontrun{
res= bubble.enrich.ms(ps= ps.ms3,dif.method = "wilcox")
}
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njau.edu.cn}
}
