% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_rmt_scale_free2.R
\name{calculate_nnsd_simple}
\alias{calculate_nnsd_simple}
\title{Simplified NNSD calculation - Alternative method}
\usage{
calculate_nnsd_simple(eigenvals)
}
\description{
Simplified NNSD calculation - Alternative method
}
