% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cir_barplot.micro.R
\name{cir_barplot.micro}
\alias{cir_barplot.micro}
\title{Generate Circular Dendrogram with Bar Plots for Microbial Data}
\usage{
cir_barplot.micro(
  ps = ps,
  Top = 15,
  dist = "bray",
  cuttree = 3,
  hcluter_method = "complete"
)
}
\arguments{
\item{ps}{A `phyloseq` object containing microbiome data.}

\item{Top}{An integer specifying the number of top taxa to display. Taxa not in the top are grouped as `"Other"`. Default is `15`.}

\item{dist}{A character string specifying the distance metric for clustering. Default is `"bray"`.}

\item{cuttree}{An integer specifying the number of clusters to cut the dendrogram into. Default is `3`.}

\item{hcluter_method}{A character string specifying the hierarchical clustering method. Default is `"complete"`.}
}
\value{
A list containing the following elements:
\describe{
  \item{Circular Dendrogram with Bar Plots}{A `ggtree` circular dendrogram with bar plots showing relative abundances of taxa.}
  \item{Processed Data}{A data frame containing the processed data used for plotting, including taxonomic annotations and abundances.}
}
}
\description{
This function generates a circular hierarchical clustering dendrogram combined with bar plots of microbial composition.
It uses taxonomic relative abundances at the Phylum level (default) and allows grouping of samples into clusters.
}
\details{
The function performs the following steps:
\itemize{
  \item Standardizes the input microbiome data using `ggClusterNet::scale_micro`.
  \item Calculates a distance matrix using the specified method (`dist`) and performs hierarchical clustering.
  \item Cuts the clustering dendrogram into the specified number of clusters (`cuttree`) and combines cluster information with sample metadata.
  \item Groups microbial taxa at the Phylum level and calculates relative abundances.
  \item Selects the top `Top` taxa by abundance, grouping the rest as `"Other"`.
  \item Combines the circular dendrogram with bar plots showing the relative abundances of taxa for each sample.
}
}
\examples{
\dontrun{
res = cir_barplot.micro(ps = ps.16s, Top = 15, dist = "bray", cuttree = 3, hcluter_method = "complete")
p17 = res[[1]]
dat = res[[2]]
}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Peng-Hao Xie \email{2019103106@njau.edu.cn}
}
