% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_associated_metabolites.R
\name{clean_data_for_ml}
\alias{clean_data_for_ml}
\title{Clean Data for Machine Learning}
\usage{
clean_data_for_ml(metabolome_data, microbe_abundance)
}
\arguments{
\item{metabolome_data}{Matrix of metabolome data}

\item{microbe_abundance}{Numeric vector of microbe abundance}
}
\value{
Cleaned metabolome data matrix
}
\description{
Removes NA values and zero variance features
}
\keyword{internal}
