% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluMicro.bar.metm.R
\name{cluMicro.bar.metm}
\alias{cluMicro.bar.metm}
\title{Generate Clustered Bar Plots for Microbial Data}
\usage{
cluMicro.bar.metm(
  dist = "bray",
  otu = NULL,
  tax = NULL,
  map = NULL,
  tree = NULL,
  j = "Phylum",
  ps = ps,
  rep = 6,
  Top = 10,
  tran = TRUE,
  hcluter_method = "complete",
  Group = "Group",
  cuttree = 3
)
}
\arguments{
\item{dist}{A character string specifying the distance method for clustering. Default is `"bray"`.}

\item{otu}{A data frame containing OTU data. If `NULL`, the `ps` object is used.}

\item{tax}{A data frame containing taxonomy data. If `NULL`, the `ps` object is used.}

\item{map}{A data frame containing sample metadata. If `NULL`, the `ps` object is used.}

\item{tree}{A phylogenetic tree object. If `NULL`, the `ps` object is used.}

\item{j}{A character string specifying the taxonomic rank to analyze (e.g., `"Phylum"`, `"Genus"`). Default is `"Phylum"`.}

\item{ps}{A `phyloseq` object containing microbiome data.}

\item{rep}{An integer specifying the number of replicates. Default is `6`.}

\item{Top}{An integer specifying the number of top taxa to display. Taxa not in the top are grouped as `"Other"`. Default is `10`.}

\item{tran}{Logical. If `TRUE`, transforms data to relative abundances. Default is `TRUE`.}

\item{hcluter_method}{A character string specifying the hierarchical clustering method. Default is `"complete"`.}

\item{Group}{A character string specifying the grouping variable in the sample metadata. Default is `"Group"`.}

\item{cuttree}{An integer specifying the number of clusters to cut the dendrogram into. Default is `3`.}
}
\value{
A list containing the following elements:
\describe{
  \item{A `ggtree` dendrogram showing the clustering tree at the sample level.}
  \item{A stacked bar plot combined with the dendrogram at the sample level.}
  \item{A `ggtree` dendrogram showing the clustering tree at the group level.}
  \item{A stacked bar plot combined with the dendrogram at the group level.}
  \item{A data frame with microbial composition data, including relative abundances and taxonomic annotations.}
}
}
\description{
This function performs hierarchical clustering on microbial data and generates stacked bar plots and dendrograms to visualize the clustering results. The microbial data can be grouped at a specified taxonomic rank and converted to relative abundances for visualization.
}
\details{
This function:
\itemize{
  \item Standardizes the microbial data to relative abundances if `tran = TRUE`.
  \item Performs hierarchical clustering on both sample-level and group-level data using the specified distance and clustering methods.
  \item Combines the clustering dendrograms with stacked bar plots of microbial composition at the specified taxonomic rank (`j`).
  \item Groups microbial taxa not in the top `Top` by abundance as `"Other"`.
}

The function outputs plots and processed data to explore microbial community composition and clustering relationships.
}
\examples{
\dontrun{result <-  cluMicro.bar.metm (dist = "bray",ps = ps.16s,j = "Genus",Top = 10, # 提取丰度前十的物种注释tran = TRUE, # 转化为相对丰度值hcluter_method = "complete",Group = "Group",cuttree = length(unique(phyloseq::sample_data(ps)$Group)))
result[[1]]
p5_2 <- result[[2]]
p5_3 <- result[[3]]
p5_4 <- result[[4]]
clubardata <- result[[5]]
}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
