% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_plot.ms.R
\name{cluster_plot.ms}
\alias{cluster_plot.ms}
\title{Create a cluster plot with tree visualization}
\usage{
cluster_plot.ms(
  ps = ps,
  hcluter_method = "complete",
  dist = "bray",
  cuttree = gnum,
  row_cluster = T,
  col_cluster = T
)
}
\arguments{
\item{ps}{A phyloseq object containing metabolite data and sample metadata.}

\item{hcluter_method}{The method for hierarchical clustering. Default is "complete".}

\item{dist}{The distance metric to be used. Default is "bray".}

\item{cuttree}{The number of clusters to cut the tree into.}

\item{row_cluster}{Logical flag indicating whether to perform row clustering. Default is TRUE.}

\item{col_cluster}{Logical flag indicating whether to perform column clustering. Default is TRUE.}
}
\value{
A list containing the following components:
\item{p0}{Cluster plot with tree visualization.}
\item{p1}{Heatmap showing relative abundance with row clustering.}
\item{p2}{Heatmap showing relative abundance with sample clustering.}
\item{tem}{Processed data for heatmap plotting.}
}
\description{
The cluster_plot.ms function generates a cluster plot with tree visualization based on OTU data and sample metadata. It includes visualization of hierarchical clustering trees and relative abundance heatmaps.
}
\examples{
# Create a cluster plot with tree visualization
cluster_plot.ms(ps = ps_object, hcluter_method = "complete", dist = "bray", cuttree = 5)

}
