% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNPS.network.R
\name{CNPS.network}
\alias{CNPS.network}
\title{Construct Co-occurrence Network for CNPS Data}
\usage{
CNPS.network(ps = ps, dat = dat, id.0 = "C")
}
\arguments{
\item{ps}{A phyloseq object containing OTU/ASV abundance data and taxonomy information}

\item{dat}{A data frame containing additional metadata or gene annotation information}

\item{id.0}{Character string specifying the prefix for node IDs in the network (default: "C")}
}
\value{
A list containing network analysis results. Typically includes:
\itemize{
  \item Network graph object
  \item Node and edge attributes
  \item Network visualization
}
}
\description{
This function constructs a co-occurrence network from CNPS gene
}
\examples{
\dontrun{
# Example usage:
dat = db.cnps
network_result <- CNPS.network(
    ps = ps.trans,
    dat = dat,
    id.0 = "G")

# Visualize the network
plot(network_result$graph)
}

}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njau.edu.cn}
}
