% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNPS.network2.R
\name{CNPS.network2}
\alias{CNPS.network2}
\title{CNPS Gene Network Analysis for Phyloseq Object}
\usage{
CNPS.network2(ps = ps.kegg, id.0 = "C")
}
\arguments{
\item{ps}{A `phyloseq` object (default is `ps.kegg`) containing OTU data to be mapped to CNPS gene groups.}

\item{id.0}{A string representing the CNPS gene group to focus on (default is `"C"`).}
}
\value{
A list containing:
  \itemize{
    \item A ggplot object (`p1`) representing the CNPS gene network plot.
    \item A data frame (`nodes`) containing the nodes of the network.
    \item A data frame (`edge`) containing the edges of the network.
  }
}
\description{
This function processes a `phyloseq` object by mapping the OTU data to CNPS gene groups,
performs correlation-based network analysis, and visualizes the network using ggplot2.
It returns a list containing the network plot, nodes, and edges.
}
\details{
This function performs the following steps:
\itemize{
  \item Filters the CNPS gene data to focus on a specific gene group (`id.0`).
  \item Extracts OTU data from the `phyloseq` object and transforms it.
  \item Subsets the OTU data for the specified CNPS gene group.
  \item Performs correlation-based analysis on the OTU data using Pearson's method.
  \item Creates a network graph using the `igraph` package based on the correlation matrix.
  \item Identifies and visualizes the network using ggplot2, with nodes representing OTUs and edges representing correlations.
}
}
\examples{
# Assuming you have a phyloseq object `ps.kegg`
network_results <-CNPS.network2(ps = ps.kegg,id.0 = "C")
network_plot <- network_results[[1]]

}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
