% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnps_ps.R
\name{cnps_ps}
\alias{cnps_ps}
\title{CNPS Gene Data for Phyloseq Object}
\usage{
cnps_ps(psko = ps.kegg)
}
\arguments{
\item{psko}{A `phyloseq` object (default is `ps.kegg`) containing OTU data to be mapped to CNPS gene information.}
}
\value{
A `phyloseq` object containing the OTU data, taxonomic data, and sample data, with CNPS gene information included.
}
\description{
This function processes a `phyloseq` object by mapping the OTU data to CNPS gene information,
and returns a new `phyloseq` object with the CNPS gene data included.
}
\details{
This function performs the following steps:
\itemize{
  \item Loads the CNPS gene data from `EasyMultiOmics.db::db.cnps`.
  \item Extracts OTU data from the `phyloseq` object and transforms it into a data frame.
  \item Joins the OTU data with the CNPS gene data using the KO identifier.
  \item Creates a unique ID for each sample by combining the CNPS gene group and KO number.
  \item Constructs a new `phyloseq` object, including the OTU table, taxonomy, and sample data, with CNPS gene information.
}
}
\examples{
# Assuming you have a phyloseq object `ps.kegg`
ps_cnps<- cnps_ps(ps.kegg)

}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
