% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_associated_metabolites.R
\name{comprehensive_correlation}
\alias{comprehensive_correlation}
\title{Comprehensive Correlation Analysis}
\usage{
comprehensive_correlation(microbe_vec, metabolome_mat)
}
\arguments{
\item{microbe_vec}{Numeric vector of microbe abundance}

\item{metabolome_mat}{Matrix of metabolome data}
}
\value{
Data frame with correlation results
}
\description{
Calculates multiple correlation metrics between microbe and metabolites
}
