% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_rmt_scale_free2.R
\name{compute_correlation_matrix}
\alias{compute_correlation_matrix}
\title{Compute correlation matrix}
\usage{
compute_correlation_matrix(data, method)
}
\description{
Compute correlation matrix
}
