% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_amplicon_result_dir.R,
%   R/create_omics_result_dir_auto.R
\name{create_omics_result_dir_auto}
\alias{create_omics_result_dir_auto}
\title{Automatically create result directory based on sequencing type}
\usage{
create_omics_result_dir_auto(
  ps,
  base_dir = "./result",
  include_time = FALSE,
  prefix = NULL,
  verbose = TRUE,
  force_data_type = NULL
)

create_omics_result_dir_auto(
  ps,
  base_dir = "./result",
  include_time = FALSE,
  prefix = NULL,
  verbose = TRUE,
  force_data_type = NULL
)
}
\arguments{
\item{ps}{A phyloseq object.}

\item{base_dir}{Base directory for all results (default: \code{"./result"}).}

\item{include_time}{Logical; if \code{TRUE}, append time (HHMMSS) to
folder name.}

\item{verbose}{Logical; if \code{TRUE}, print inferred sequencing type.}

\item{metagenome_prefix}{Prefix for metagenome folders
(default: \code{"metagenome"}).}

\item{amplicon_prefix}{Prefix for amplicon folders
(default: \code{"amplicon"}).}
}
\value{
A character scalar giving the path of the created directory.

A character scalar giving the path of the created directory.
}
\description{
If the phyloseq object is inferred as metagenome, a folder like
\code{"metagenome_YYYYMMDD"} (or with time) will be created.
If inferred as amplicon (or ambiguous), it will call
\code{create_amplicon_result_dir()} to create a folder such as
\code{"amplicon_multi_Bac60-Fun40_YYYYMMDD"}.

If the phyloseq object is inferred as metagenome, a folder like
\code{"metagenome_YYYYMMDD"} (or with time) will be created.
If inferred as amplicon (or ambiguous), it will call
\code{create_amplicon_result_dir()} to create a folder such as
\code{"amplicon_multi_Bac60-Fun40_YYYYMMDD"}.
}
