% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_associated_metabolites.R
\name{cross_omics_normalization}
\alias{cross_omics_normalization}
\title{Cross-omics Data Normalization}
\usage{
cross_omics_normalization(microbe_abundance, metabolome_data)
}
\arguments{
\item{microbe_abundance}{Numeric vector of microbe abundance}

\item{metabolome_data}{Matrix of metabolome data}
}
\value{
A list containing normalized data
}
\description{
Normalizes microbiome and metabolome data for cross-omics comparison
}
