% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dacomp.trans.R
\name{dacomp.trans}
\alias{dacomp.trans}
\title{Perform differential analysis using DACOMP}
\usage{
dacomp.trans(ps = ps0, sd = 0.6, q_BH = 0.1, q_DSFDR = 0.1)
}
\arguments{
\item{ps}{A phyloseq format file used as an alternative for the input containing transcriptome functional composition table, tax, and sample metadata.}

\item{sd}{Standard deviation threshold for selecting reference taxa.}

\item{q_BH}{Benjamini-Hochberg correction threshold.}

\item{q_DSFDR}{DS-FDR correction threshold.}
}
\value{
A list containing the results of significant differences in genes abundance between different groups.
}
\description{
This function is used to analyze transcriptome functional composition
 data  by conducting DACOMP analysis to detect significant differences in genes
 abundance between different groups.
}
\examples{
\dontrun{
data(ps.trans)
ps = ps.trans\%>\% filter_taxa(function(x) sum(x ) > 5 , TRUE)
dat = dacomp.trans (ps =  ps\%>\% filter_OTU_ps(50) ,sd = 0.6,q_BH =  0.1,q_DSFDR =0.1)
dat1 = dat$WT_OE
head(dat1)
dat2 = dat$WT_KO
head(dat2)
dat3 = dat$OE_KO
head(dat3)
}
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}.
}
