% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisiontree.micro.R
\name{decisiontree.micro}
\alias{decisiontree.micro}
\title{Decisiontree model screening of characteristic microorganisms}
\usage{
decisiontree.micro(ps = ps, top = 20, seed = 6358, k = 5)
}
\arguments{
\item{ps}{A phyloseq format file used as an alternative for the input containing otu, tax, and map.}

\item{top}{The top microorganisms to consider.}

\item{seed}{The random seed for reproducibility.}

\item{k}{The number of folds for cross-validation.}
}
\value{
A list object including the following components:
\item{Accuracy}{The average accuracy of the Decisiontree model.}
\item{Importance}{A data frame showing the feature importance ranked in descending order.}
}
\description{
Decisiontree, one of the machine learning methods, was used to screen for characteristic
microorganisms, and the model was evaluated using k-fold cross-validation.
}
\examples{
library(dplyr)
library(ggClusterNet)
library(caret)
library(rpart)
res = decisiontree.micro(ps=ps, top = 20, seed = 1010, k = 5)
accuracy = res[[1]]
accuracy
importance = res[[2]]
importance
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
