% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_associated_metabolites.R
\name{differential_analysis}
\alias{differential_analysis}
\title{Differential Analysis}
\usage{
differential_analysis(microbe_abundance, metabolome_data, groups = NULL)
}
\arguments{
\item{microbe_abundance}{Numeric vector of microbe abundance}

\item{metabolome_data}{Matrix of metabolome data}

\item{groups}{Optional grouping vector}
}
\value{
Data frame with differential analysis results
}
\description{
Performs differential expression analysis based on microbe abundance
}
