% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ven.network.trans.R
\name{div_network2}
\alias{div_network2}
\title{Generate a Circular transcriptome functional Network}
\usage{
div_network2(ps, group = "Group", flour = TRUE, N = 0.5)
}
\arguments{
\item{ps}{A phyloseq format file used as an alternative for the input containing transcriptome functional composition table, tax, and sample metadata.}

\item{N}{Numeric. A threshold for selecting nodes in the network. Default is `0.5`.}

\item{fill}{A character string specifying the taxonomic rank used to color the nodes. Default is `"Phylum"`.}
}
\value{
A list containing:
\describe{
  \item{plot}{A `ggplot2` object representing the transcriptome functional network.}
  \item{plotdata}{A data frame containing node coordinates and associated metadata.}
}
}
\description{
This function generates a circular transcriptome functional composition based on transcriptome functional abundance data in a `phyloseq` object.
It uses the `ggClusterNet` package to compute relationships between genes
and visualizes the network with sample-level metadata.
}
\details{
This function builds a transcriptome functional network using a circular layout. OTUs are placed hierarchically based on their
abundance and relationships with samples. The edges represent connections between genes and sample groups,
while nodes represent transcriptome functional taxa or samples.
}
\examples{
library(ggClusterNet)
data(ps.trans)
ps = ps.trans\%>\% filter_taxa(function(x) sum(x ) > 5 , TRUE)
result = ven.network.trans(ps = ps,N = 0.5,fill = "KO_id")
p14  = result[[1]]
p14
dat = result[[2]]
head(dat)
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Peng-Hao Xie \email{2019103106@njqu.edu.cn}
The function includes several steps:
\itemize{
  \item Calculates the transcriptome functional relationships using `ggClusterNet::div_network`.
  \item Uses a progressive circle packing algorithm to position genes and groups.
  \item Combines abundance data and taxonomic annotations for visualization.
  \item Generates a `ggplot2` plot showing the network with colored nodes and edges.
}
}
