% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_minimal_change_units.R
\name{.calculate_consistency}
\alias{.calculate_consistency}
\title{Calculate Consistency and Direction}
\usage{
.calculate_consistency(otu_matrix, groups, group_levels)
}
\description{
Calculate Consistency and Direction
}
\keyword{internal}
