% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_minimal_change_units.R
\name{.check_upper_consistency}
\alias{.check_upper_consistency}
\title{Check Upper Level Consistency}
\usage{
.check_upper_consistency(
  tax_agg,
  i,
  tax_table_mat,
  otu_clr,
  groups,
  group_levels,
  levels,
  level_idx,
  consist_thresh,
  p_thresh,
  min_sub_units
)
}
\description{
Check Upper Level Consistency
}
\keyword{internal}
