% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_minimal_change_units.R
\name{.create_empty_result}
\alias{.create_empty_result}
\title{Create Empty Result Data Frame}
\usage{
.create_empty_result()
}
\description{
Create Empty Result Data Frame
}
\keyword{internal}
