% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_minimal_change_units.R
\name{.extract_full_taxonomy}
\alias{.extract_full_taxonomy}
\title{Extract Full Taxonomy from Phyloseq Object}
\usage{
.extract_full_taxonomy(results, physeq)
}
\description{
Extract Full Taxonomy from Phyloseq Object
}
\keyword{internal}
