% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_minimal_change_units.R
\name{.plot_phylogenetic_tree}
\alias{.plot_phylogenetic_tree}
\title{Create Phylogenetic Tree Plot}
\usage{
.plot_phylogenetic_tree(results, physeq, title, interactive)
}
\description{
Create Phylogenetic Tree Plot
}
\keyword{internal}
