% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_minimal_change_units.R
\name{.plot_sunburst}
\alias{.plot_sunburst}
\title{Create Sunburst Plot}
\usage{
.plot_sunburst(results, physeq, title, interactive)
}
\description{
Create Sunburst Plot
}
\keyword{internal}
