% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_minimal_change_units.R
\name{.prepare_sunburst_data}
\alias{.prepare_sunburst_data}
\title{Prepare Sunburst Data}
\usage{
.prepare_sunburst_data(results, physeq)
}
\description{
Prepare Sunburst Data
}
\keyword{internal}
