% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_minimal_change_units.R
\name{.prepare_tree_data}
\alias{.prepare_tree_data}
\title{Prepare Tree Data Structure}
\usage{
.prepare_tree_data(results, physeq)
}
\description{
Prepare Tree Data Structure
}
\keyword{internal}
