% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_minimal_change_units.R
\name{.prepare_treemap_data}
\alias{.prepare_treemap_data}
\title{Prepare Treemap Data}
\usage{
.prepare_treemap_data(results, physeq)
}
\description{
Prepare Treemap Data
}
\keyword{internal}
