% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_minimal_change_units.R
\name{.select_minimal_units}
\alias{.select_minimal_units}
\title{Select Minimal Units}
\usage{
.select_minimal_units(results_df, levels, verbose)
}
\description{
Select Minimal Units
}
\keyword{internal}
