% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_minimal_change_units.R
\name{.shorten_taxon_names}
\alias{.shorten_taxon_names}
\title{Shorten Taxon Names for Plotting}
\usage{
.shorten_taxon_names(taxon_names, max_length = 30)
}
\description{
Shorten Taxon Names for Plotting
}
\keyword{internal}
