% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_minimal_change_units.R
\name{.validate_inputs}
\alias{.validate_inputs}
\title{Validate Input Parameters}
\usage{
.validate_inputs(
  physeq,
  group_col,
  levels,
  consist_thresh,
  p_thresh,
  abund_thresh,
  min_sub_units
)
}
\description{
Validate Input Parameters
}
\keyword{internal}
