% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_Manhattan.metm.R
\name{edge_Manhattan.metm}
\alias{edge_Manhattan.metm}
\title{Generate Manhattan Plots for Differential Abundance Analysis}
\usage{
edge_Manhattan.metm(ps = ps, pvalue = 0.05, lfc = 0)
}
\arguments{
\item{ps}{A `phyloseq` object containing microbiome data (OTU table, taxonomic table, and sample metadata).}

\item{pvalue}{A numeric value specifying the significance threshold for adjusted p-values (FDR). Default is `0.05`.}

\item{lfc}{A numeric value specifying the log fold change threshold for significance. Default is `0`.}
}
\value{
A list of Manhattan plots, one for each pairwise group comparison, showing:
\itemize{
  \item Log-transformed p-values (`-log(p)`) on the y-axis.
  \item OTUs (taxa) on the x-axis, arranged by Phylum.
  \item Different shapes and colors indicating significant enrichment, depletion, or nonsignificance.
  \item Phylum-specific colors and point sizes proportional to log-transformed counts per million (logCPM).
}
}
\description{
The `edge_Manhattan.metm` function performs pairwise differential abundance analysis using the `edgeR` package.
It visualizes the results for each comparison in Manhattan plots, highlighting enriched, depleted, and nonsignificant taxa
with taxonomic annotations.
}
\details{
The function performs the following steps:
\itemize{
  \item Extracts OTU and sample data from the `phyloseq` object.
  \item Normalizes the OTU counts using the `edgeR` TMM method.
  \item Performs pairwise group comparisons using contrasts in a generalized linear model (GLM) framework.
  \item Identifies significantly enriched and depleted OTUs based on FDR-adjusted p-values and log fold changes.
  \item Annotates OTUs with taxonomic information.
  \item Generates Manhattan plots for each pairwise comparison.
}

Top phyla are highlighted, while taxa from less abundant phyla are categorized as "Low Abundance."
Each Manhattan plot contains a horizontal line representing the minimum significant FDR value.
}
\examples{
\dontrun{
res = edge_Manhattan.metm(ps = ps.16s\%>\% ggClusterNet::filter_OTU_ps(500),pvalue = 0.05,lfc = 0)
 }

}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
