% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgerSuper.metf.R
\name{EdgerSuper.metf}
\alias{EdgerSuper.metf}
\title{Functional expression difference analysis}
\usage{
EdgerSuper.metf(
  otu = NULL,
  tax = NULL,
  map = NULL,
  tree = NULL,
  ps = NULL,
  j = "meta",
  group = "Group",
  pvalue = 0.05,
  artGroup = NULL
)
}
\arguments{
\item{otu}{A matrix of OTU (Operational Taxonomic Units) data.}

\item{tax}{A taxonomy table associated with the OTU data.}

\item{map}{A sample metadata table.}

\item{tree}{A phylogenetic tree.}

\item{ps}{A phloseq format file used as an alternative for the input containing otu, tax, and map.}

\item{j}{The rank or type of analysis ("meta", "OTU", "gene", "Kingdom","Phylum","Class","Order","Family","Genus","Species"or taxonomy ranks).}

\item{group}{The grouping variable in the metadata table.}

\item{pvalue}{The significance threshold for differential expression.}

\item{artGroup}{An optional matrix of alternative group combinations}
}
\value{
A list object including the following components:
\item{plot_list}{A list of volcano plots containing the results of the analysis that are expressed differently between all selected groups.}
\item{x}{A combined data table representing the results of the differential expression analysis for all selected groups.}
}
\description{
This function conducts Functional expression difference analysis using
DESeq2 and generates volcano plots based on KEGG data inputs.
}
\examples{
ps.kegg = ps.kegg \%>\% filter_OTU_ps(Top = 1000)
tax =ps.kegg \%>\% tax_table()
colnames(tax)[3] = "KOid"
tax_table(ps.kegg) =tax
res = EdgerSuper.metf (ps = ps.kegg,group  = "Group",artGroup = NULL)
dat = res[[2]]
dat
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
