% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgerSuper.trans.R
\name{EdgerSuper.trans}
\alias{EdgerSuper.trans}
\title{Differential Abundance Analysis Using EdgeR for Metagenome Functional Data}
\usage{
EdgerSuper.trans(
  otu = NULL,
  tax = NULL,
  map = NULL,
  tree = NULL,
  ps = NULL,
  group = "Group",
  pvalue = 0.05,
  lfc = 0,
  artGroup = NULL,
  method = "TMM",
  j = "meta"
)
}
\arguments{
\item{otu}{A data frame containing metagenome functional counts. Optional if `ps` is provided.}

\item{tax}{A data frame containing taxonomic annotations. Optional if `ps` is provided.}

\item{map}{A data frame containing sample metadata. Optional if `ps` is provided.}

\item{tree}{A phylogenetic tree. Optional if `ps` is provided.}

\item{ps}{A `phyloseq` object containing metagenome functional data. If provided, overrides `otu`, `tax`, `map`, and `tree`.}

\item{group}{A string specifying the grouping variable in the sample metadata. Default is `"Group"`.}

\item{pvalue}{A numeric value specifying the significance threshold for adjusted p-values. Default is `0.05`.}

\item{lfc}{A numeric value specifying the log2 fold change cutoff. Default is `0`.}

\item{artGroup}{A matrix specifying custom group comparisons. Optional.}

\item{method}{A string specifying the normalization method for EdgeR. Default is `"TMM"`.}

\item{j}{A string or integer specifying the taxonomic rank to perform the analysis on. Metagenome functional data should select "meta".}
}
\value{
A list containing:
\describe{
  \item{Plots}{A list of volcano plots for each group comparison, showing log fold changes and p-values.}
  \item{Results}{A data frame containing differential abundance results, including normalized abundances and taxonomic annotations.}
}
}
\description{
This function performs differential abundance analysis for metagenome functional data using the `EdgeR` package.
It identifies differentially abundant taxa between groups in a `phyloseq` object, visualizes the results
with volcano plots, and outputs normalized abundance data with taxonomic annotations.
}
\details{
The function performs the following steps:
\itemize{
  \item Prepares metagenome functional data, taxonomic annotations, and metadata from a `phyloseq` object or individual input files.
  \item Aggregates data to the specified taxonomic rank (if applicable).
  \item Normalizes metagenome functional counts using the specified normalization method (`"TMM"` by default).
  \item Constructs a model matrix for group comparisons and estimates dispersions using EdgeR's generalized linear model (GLM) framework.
  \item Identifies differentially abundant taxa based on the specified `pvalue` and `lfc` thresholds.
  \item Generates volcano plots for each group comparison, highlighting enriched and depleted taxa.
  \item Appends taxonomic annotations and normalized abundances to the differential abundance results.
}

The function supports both predefined and custom group comparisons (`artGroup`) and provides detailed visualizations for the results.
}
\examples{
\dontrun{
res = EdgerSuper.trans(ps = ps.kegg \%>\%
                       ggClusterNet::filter_OTU_ps(Top=500),
                       group = "Group",
                       artGroup = NULL,
                       j = "meta")
p25.1 = res[[1]][1]
p25.1
p25.2 = res[[1]][2]
p25.2
p25.3 = res[[1]][3]
p25.3
dat = res[[2]]
head(dat)
}

}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Peng-Hao Xie \email{2019103106@njau.edu.cn}
}
