% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEAST.micro.R
\name{FEAST.micro}
\alias{FEAST.micro}
\title{FEAST Microbial traceability analysis}
\usage{
FEAST.micro(
  otu = otutab,
  map = metadata,
  ps = NULL,
  group = "Group",
  sinkG = "WT",
  sourceG = c("KO", "OE"),
  rep = NULL
)
}
\arguments{
\item{otu}{OTU/ASV table;}

\item{map}{Sample metadata;}

\item{group}{group ID;}

\item{sinkG}{object of sink group}

\item{sourceG}{object of source group}
}
\value{
a table
}
\description{
Input otutab, metadata or phyloseq object; ; output a table object.
}
\details{
By default, input phyloseq object include metadata and otutab
}
\examples{
result = FEAST.micro(ps = ps.16s,group = "Group",sinkG = "OE",sourceG = c("WT","KO"))
FEAST(otu = otu,map = map,group = "Group",sinkG = "WT",sourceG = c("KO","OE"))
FEAST(ps =ps,group = "Group",sinkG = "WT",sourceG = c("KO","OE"))

}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
