% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_metabolite_selection.R
\name{find_natural_breaks}
\alias{find_natural_breaks}
\title{Find Natural Breaks in Score Distribution}
\usage{
find_natural_breaks(scores)
}
\arguments{
\item{scores}{Numeric vector of metabolite scores (sorted descending)}
}
\value{
Integer, index of the first natural break
}
\description{
Identifies natural groupings in metabolite scores using Jenks natural breaks
or standard deviation-based thresholding.
}
\keyword{internal}
