% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggflower.metf.R
\name{ggflower.metf}
\alias{ggflower.metf}
\title{Create a circular flower plot based on metagenome functional composition data}
\usage{
ggflower.metf(
  otu = NULL,
  tax = NULL,
  map = NULL,
  ps = NULL,
  group = "Group",
  rep = 6,
  m1 = 2,
  start = 1,
  a = 0.2,
  b = 1,
  lab.leaf = 1,
  col.cir = "yellow",
  a.cir = 0.5,
  b.cir = 0.5,
  m1.cir = 2,
  N = 0.5
)
}
\arguments{
\item{otu}{Metagenome functional composition table.}

\item{tax}{Metagenome functional classification table.}

\item{map}{Sample metadata.}

\item{ps}{A phyloseq format file used as an alternative for the input containing metagenome functional composition table, tax, and sample metadata.}

\item{group}{Column name for groupID in map table(sample metadata).}

\item{rep}{Repeat number of metagenome functional composition data.}

\item{m1}{Petal shape, square to round to prismatic, the value gradually decreases.}

\item{start}{The rotation angle of the petals, the greater the value, the greater the angle.}

\item{a}{The width of the petals.}

\item{b}{Distance from petal to center.}

\item{lab.leaf}{The distance from the label to the center of the circle.}

\item{col.cir}{Center color.}
}
\value{
A list containing the following components:
\item{p}{Circular flower plot visualizing the presence or absence of genes across sample groups.}
\item{ven2}{Processed genes data frame used for plotting.}
}
\description{
The ggflower.metf function generates a circular flower plot based on metagenome functional composition data.
 The plot visualizes the presence or absence of genes
 across different sample groups in a circular layout.
}
\examples{
library(phyloseq)
library(ggplot2)
ps =ps.kegg \%>\% filter_OTU_ps(Top = 1000)
res <- ggflower.metf(ps= ps ,group = "Group",start = 1,
m1 = 1,a = 0.2, b = 1,lab.leaf = 1,col.cir = "yellow",N = 0.5 )
p14 = res[[1]]
p14
dat = res[[2]]
dat
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
