% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggflower.ms.R
\name{ggflower.ms}
\alias{ggflower.ms}
\title{Create a circular flower plot based on metabolite data}
\usage{
ggflower.ms(
  otu = NULL,
  tax = NULL,
  map = NULL,
  ps = NULL,
  group = "Group",
  rep = 6,
  m1 = 2,
  start = 1,
  a = 0.2,
  b = 1,
  lab.leaf = 1,
  col.cir = "yellow",
  a.cir = 0.5,
  b.cir = 0.5,
  m1.cir = 2,
  N = 0.5
)
}
\arguments{
\item{otu}{otu table of metabolite,data.frame.}

\item{tax}{taxonmy table of metabolite,data.frame.}

\item{map}{Sample metadata.}

\item{ps}{A phyloseq format file used as an alternative for the input containing metabolite composition table,
metabolite classification table, and sample metadata.}

\item{group}{column name for groupID in map table(sample metadata).}

\item{rep}{repeat number of microbial data.}

\item{m1}{Petal shape, square to round to prismatic, the value gradually decreases.}

\item{start}{The rotation angle of the petals, the greater the value, the greater the angle.}

\item{a}{The width of the petals.}

\item{b}{Distance from petal to center.}

\item{lab.leaf}{The distance from the label to the center of the circle.}

\item{col.cir}{Center color.}
}
\value{
A list containing the following components:
\item{p}{Circular flower plot visualizing the presence or absence of metabolites across sample groups.}
\item{ven2}{Processed metabolite data frame used for plotting.}
}
\description{
The ggflower.ms function generates a circular flower plot based on metabolite data.
 The plot visualizes the presence or absence of metabolites
 across different sample groups in a circular layout.
}
\examples{
library(phyloseq)
library(ggplot2)
res <- ggflower.ms(ps= ps.ms ,rep = 1,group = "Group",start = 1,
m1 = 1,a = 0.2, b = 1,lab.leaf = 1col.cir = "yellow",N = 0.5 )
p14 = res[[1]]
p14
dat = res[[2]]
dat
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
