% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggVen.Upset.micro.R
\name{ggVen.Upset.micro}
\alias{ggVen.Upset.micro}
\title{Plotting Ven and Upset plot for each group}
\usage{
ggVen.Upset.micro(
  otu = NULL,
  tax = NULL,
  map = NULL,
  tree = NULL,
  ps = NULL,
  group = "Group",
  N = 0.5
)
}
\arguments{
\item{group}{column name for groupID.}

\item{otutab}{rarefied OTU table, typical output of usearch -otutab_norm or -otutab_rare,}

\item{metadata}{matrix or dataframe, including sampleID and groupID;}

\item{rep}{repeat number of each group}
}
\value{
base plot object.
}
\description{
Input otutab and metadata
VennDiagram::Ven plot
UpSetR::Upset plot
}
\examples{
# Set four parameters: otutab, metadata, group and rep
VenUpset(otu = otutab,map = metadata,group = "genotype",rep = 6)
}
\references{
Zhang, J., Zhang, N., Liu, Y.X., Zhang, X., Hu, B., Qin, Y., Xu, H., Wang, H., Guo, X., Qian, J., et al. (2018).
Root microbiota shift in rice correlates with resident time in the field and developmental stage.
Sci China Life Sci 61, DOI: \url{https://doi.org/10.1007/s11427-018-9284-4}
}
\seealso{
Ven-Upset
}
\author{
Contact: Yong-Xin Liu \email{metagenome@126.com}
}
