% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.trans.R
\name{glm.trans}
\alias{glm.trans}
\title{Glm model screening of characteristic genes}
\usage{
glm.trans(ps = ps, k = 5)
}
\arguments{
\item{ps}{A phyloseq format file used as an alternative for the input containing transcriptome functional composition table,
transcriptome functional classification table, and sample metadata.}

\item{k}{The number of folds for cross-validation.}
}
\value{
A list object including the following components:
\item{AUC}{The average accuracy of the glm model.}
\item{Importance}{A data frame showing the feature importance ranked in descending order.}
}
\description{
Glm, one of the machine learning methods, was used to screen for characteristic
genes, and the model was evaluated using k-fold cross-validation.
}
\examples{
library(dplyr)
library(ggClusterNet)
library(caret)
library(ROCR)
data(ps.trans)
ps =ps.trans \%>\% filter_OTU_ps(Top = 100)
pst=ps\%>\%subset_samples(ps,Group \%in\% c("KO" ,"OE"))
res <- glm.trans(ps = pst \%>\% filter_OTU_ps(50), k = 5)
AUC = res[[1]]
AUC
importance = res[[2]]
importance
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njau.edu.cn}
}
