% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_omics_result_dir_auto.R
\name{guess_phyloseq_data_type}
\alias{guess_phyloseq_data_type}
\title{Guess the data type of a phyloseq object (amplicon vs metabolome)}
\usage{
guess_phyloseq_data_type(ps)
}
\arguments{
\item{ps}{A \code{phyloseq} object.}
}
\value{
A character scalar, one of:
  \itemize{
    \item \code{"metabolome"} – looks like metabolomics data;
    \item \code{"amplicon"}   – looks like amplicon microbiome data;
    \item \code{"unknown"}    – cannot be reliably classified.
  }
}
\description{
This function inspects the taxonomy table of a phyloseq object and
tries to guess whether it represents amplicon-style data (with
biological taxonomy such as Kingdom / Phylum) or metabolomics-style
data (with HMDB / CAS / KEGG compound IDs, etc.).
}
\examples{
\dontrun{
  guess_phyloseq_data_type(ps)
}

}
