% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.ms.R
\name{heatmap.ms}
\alias{heatmap.ms}
\title{Create a heatmap with metabolites clustering and labeling}
\usage{
heatmap.ms(
  ps_rela = ps_rela,
  label = TRUE,
  col_cluster = TRUE,
  row_cluster = TRUE
)
}
\arguments{
\item{ps_rela}{A phyloseq format file used as an alternative for the input containing metabolite composition table(realtive abundance),
metabolite classification table, and sample metadata.}

\item{label}{Logical parameters indicating whether to label sample groups. Default is TRUE.}

\item{col_cluster}{Logical parameters indicating whether to perform column clustering. Default is TRUE.}

\item{row_cluster}{Logical parameters indicating whether to perform row clustering. Default is TRUE.}
}
\value{
A list containing the following components:
\item{p1}{Heatmap with row clustering and relative abundance visualization.}
\item{p2}{Bubble heatmap with sample clustering and relative abundance visualization.}
\item{pcm}{Analysis data for heatmap plotting.}
}
\description{
The heatmap.ms function generates a heatmap with clustering and labeling based on metabolites  data and sample metadata.
 It includes row and column clustering, relative abundance visualization, and sample group labeling.
}
\examples{
ps.ms_rela <- ps.ms \%>\% scale_micro(method = "rela") \%>\%
 tax_glom_wt(ranks = "Class")
 result <- heatmap.ms (ps.ms_rela,label =  FALSE,col_cluster = FALSE,row_cluster =FALSE)
 p19 <- result[[1]]
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
