% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub.gene.trans.R
\name{hub.gene.trans}
\alias{hub.gene.trans}
\title{Cluster and Visualize Gene Expression Data}
\usage{
hub.gene.trans(
  ps_pro = ps.trans,
  id = "KO_id",
  cluster.method = "mfuzz",
  cluster.num = 6
)
}
\arguments{
\item{ps_pro}{A phyloseq object containing the gene expression data}

\item{id}{Character string specifying the column name to use for gene IDs (default: "KO_id")}

\item{cluster.method}{Character string specifying the clustering method to use (default: "mfuzz")}

\item{cluster.num}{Numeric specifying the number of clusters to generate (default: 6)}
}
\value{
A list containing:
\itemize{
  \item p1 - Line plot visualization of clusters
  \item p2 - Combined heatmap and line plot visualization of clusters
  \item dat - Cluster data including cluster assignments
}
}
\description{
This function performs clustering on gene expression data and generates visualization plots.
It takes a phyloseq object, processes the data, performs clustering using specified method,
and returns cluster visualizations and data.
}
\examples{
\dontrun{
# Example usage:
result <- hub.gene.trans(ps_pro = my_phyloseq_object)
result$p1  # View line plot
result$p2  # View combined plot
cluster_data <- result$dat  # Access cluster data
}

}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Peng-Hao Xie \email{2019103106@njau.edu.cn}
}
