% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub.network.omics.R
\name{hub.network.omics}
\alias{hub.network.omics}
\title{hub.network.omics}
\usage{
hub.network.omics(cor, top = 10)
}
\arguments{
\item{cor}{A data frame representing a correlation matrix.}

\item{top}{An integer specifying how many top hub nodes to display (default: `10`).}
}
\value{
A ggplot object representing the bar plot of the top hub nodes.
}
\description{
This function calculates the hub scores from a correlation matrix, identifies the top hub nodes,
and visualizes the results in a bar plot. It uses the `igraph` package to create a graph from the
correlation data and then computes hub scores. The top `tem` hub nodes are extracted and displayed.
}
\examples{
hub_plot <- hub.network.omics(cor = correlation_matrix, top = 10)
print(hub_plot)

}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
