% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_metabolite_selection.R
\name{independent_components}
\alias{independent_components}
\title{Independent Components Analysis}
\usage{
independent_components(results)
}
\arguments{
\item{results}{Results object from find_associated_metabolites}
}
\value{
Integer, estimated number of independent components
}
\description{
Estimates number of independent metabolites based on correlation structure.
}
\keyword{internal}
