% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_metabolite_selection.R
\name{information_gain_analysis}
\alias{information_gain_analysis}
\title{Information Gain Analysis}
\usage{
information_gain_analysis(scores)
}
\arguments{
\item{scores}{Numeric vector of metabolite scores (sorted descending)}
}
\value{
Integer, optimal number of metabolites based on information gain
}
\description{
Calculates the point where adding more metabolites provides diminishing returns
based on information gain rate.
}
\keyword{internal}
