% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG_enrich.micro.R
\name{KEGG_enrich.micro}
\alias{KEGG_enrich.micro}
\title{KEGG Pathway Enrichment Analysis for Microbial Communities}
\usage{
KEGG_enrich.micro(ps = ps, dif = NULL)
}
\arguments{
\item{ps}{A `phyloseq` object containing OTU (or ASV) table, taxonomic annotations, and sample metadata.}

\item{dif}{A data frame containing differential abundance results for microbial features (e.g., genes, pathways). Must include columns with group-specific comparisons and significance labels.}
}
\value{
A list of data frames, where each data frame contains KEGG enrichment results for a specific group comparison.
}
\description{
The `KEGG_enrich.micro` function performs KEGG pathway enrichment analysis for microbial communities. It identifies enriched pathways based on differential abundance results and visualizes the results for different group comparisons.
}
\details{
The function iterates through all pairwise group comparisons, performs KEGG pathway enrichment using the `clusterProfiler::enrichKEGG` function, and returns the enrichment results for each comparison. It filters pathways based on adjusted p-value and q-value thresholds and prepares data for visualization.

Steps:
\itemize{
  \item Extracts all pairwise group comparisons from the sample metadata in the `phyloseq` object.
  \item For each comparison, identifies significantly different features (`ko` identifiers).
  \item If no significant features are found, selects the top 500 features for analysis.
  \item Runs KEGG pathway enrichment analysis using the `enrichKEGG` function.
  \item Filters and organizes enrichment results for visualization (e.g., bubble plots).
}
}
\examples{
\dontrun{
res = EdgerSuper.metf (ps = ps.kegg,group  = "Group",artGroup = NULL)
dat = res[[2]]
res2 = KEGG_enrich.micro(ps =  ps.kegg,  diffpath = diffpath,dif = dat)

}

}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
