% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG_enrich.tran.R
\name{KEGG_enrich.trans}
\alias{KEGG_enrich.trans}
\title{KEGG Pathway Enrichment Analysis with Differential Expression}
\usage{
KEGG_enrich.trans(ps = ps, method.diff = md[4])
}
\arguments{
\item{ps}{A Phyloseq object containing OTU table and sample data. The default is `ps`,
which assumes that the Phyloseq object is provided externally.}

\item{method.diff}{A string specifying the method for differential analysis. The default
is `md[4]`, which should be set appropriately based on the method you wish to use.}
}
\value{
A list containing:
\itemize{
  \item \code{plots}: A list of generated plots, including bubble and bar plots for
  KEGG pathway enrichment.
  \item \code{plotdata}: A list of data used for plotting, including enriched pathway
  information.
}
}
\description{
This function performs KEGG pathway enrichment analysis using KEGG Orthology (KO)
terms. It analyzes differential expression results from the `EdgerSuper.trans()` function
and identifies significantly enriched KEGG pathways. The function also generates
various plots to visualize the KEGG enrichment results, including bubble and bar plots.
}
\examples{
\dontrun{
result <- KEGG_enrich.trans(ps = ps.trans)
}
}
