% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan.omics.R
\name{lavaan.omics}
\alias{lavaan.omics}
\title{Multi-omics structural equation model construction--Lavaan}
\usage{
lavaan.omics(
  ps.trans = ps.trans,
  ps.ms = ps.ms,
  ps.micro = ps.micro,
  ps.meta = ps.kegg,
  filter = 0.001
)
}
\arguments{
\item{ps.trans}{Transcriptome data for analysis, a phyloseq format file used as an alternative for the input containing transcriptome functional composition table.}

\item{ps.ms}{Metabolome data for analysis, a phyloseq format file used as an alternative for the input containing metabolites composition table.}

\item{ps.micro}{Microbiome data for analysis, a phyloseq format file used as an alternative for the input containing microbial composition table.}

\item{ps.meta}{Metagenome functional data for analysis, a phyloseq format file used as an alternative for the input containing metagenome functional composition table.}

\item{filter}{Threshold for filtering low abundance data,default is 0.001.}
}
\value{
Returns a data frame containing model fit statistics.
}
\description{
This function is used to conduct Structural Equation Modeling (SEM) analysis by lavaan,
the relationship between Transcriptome,Metabolome,Microbiome and Metagenome is explored by extracting the results of the first axis of PCA analysis.
}
\examples{
\dontrun{
library(vegan)
library(phyloseq)
library(dplyr)
library(ggClusterNet)
library(lavaan)
library(semPlot)
library(EasyMultiOmics)
data(ps.trans)
data(ps.ms)
data(ps.micro)
data(ps.kegg)
res <- lavaan.omics(ps.trans=ps.trans,ps.ms=ps.ms,ps.micro=ps.micro,
ps.meta=ps.kegg,filter=0.05)
res
}
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njau.edu.cn}.
}
