% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.metf.R
\name{LDA.metf}
\alias{LDA.metf}
\title{Perform LDA analysis on metagenome functional composition data}
\usage{
LDA.metf(
  ps = ps.cs,
  group = "Group",
  Top = 100,
  p.lvl = 0.05,
  lda.lvl = 2,
  seed = 11,
  adjust.p = F
)
}
\arguments{
\item{ps}{A phyloseq format file used as an alternative for the input containing metagenome functional composition table, tax, and map.}

\item{group}{Column name for groupID in map table(sample metadata).}

\item{Top}{The top functions to convert and visualize.}

\item{p.lvl}{p-value threshold for the characteristic functions.}

\item{lda.lvl}{LDA score threshold for the characteristic functions.}

\item{seed}{The random seed for reproducibility.}

\item{adjust.p}{Logical indicating whether to adjust p-values.}
}
\value{
A list object containing the following components:
\item{lefse_lists}{Data frame containing LDA analysis results.}
\item{taxtree}{Data frame containing taxonomic information and LDA results.}
}
\description{
This function performs LDA analysis on metagenome functional composition data to
identify characteristic functions at different taxonomic levels.
}
\examples{
ps =ps.cnps\%>\% filter_OTU_ps(Top = 1000)
tablda = LDA.metf(ps = ps,
Top = 100,
p.lvl = 0.05,
lda.lvl = 1,
seed = 11,
adjust.p = F)
dat1=tablda[[1]]
head(dat1)
dat2 = tablda[[2]]
head(dat2)
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
