% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.metm.R
\name{LDA.metm}
\alias{LDA.metm}
\title{Perform LDA analysis on microbiome data}
\usage{
LDA.metm(
  ps = ps,
  group = "Group",
  Top = 100,
  p.lvl = 0.05,
  lda.lvl = 2,
  seed = 11,
  adjust.p = F
)
}
\arguments{
\item{ps}{A phyloseq format file used as an alternative for the input containing otu, tax, and map.}

\item{group}{column name for groupID in map table(sample metadata).}

\item{Top}{The top microorganisms to convert and visualize.}

\item{p.lvl}{p-value threshold for the Characteristic microorganisms.}

\item{lda.lvl}{LDA score threshold for the Characteristic microorganisms.}

\item{seed}{The random seed for reproducibility.}

\item{adjust.p}{Logical indicating whether to adjust p-values.}
}
\value{
A list object containing the following components:
\item{lefse_lists}{Data frame containing LDA analysis results.}
\item{taxtree}{Data frame containing taxonomic information and LDA results.}
}
\description{
This function performs LDA analysis on microbiome data to
identify Characteristic microorganisms at different taxonomic levels.
}
\examples{
tablda = LDA.metm(ps = ps.16s,
Top = 100,
p.lvl = 0.05,
lda.lvl = 1,
seed = 11,
adjust.p = F)
dat1=tablda[[1]]
head(dat1)
dat2 = tablda[[2]]
head(dat2)
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
