% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.ms.R
\name{LDA.ms}
\alias{LDA.ms}
\title{Linear Discriminant Analysis (LDA) for Metabolomic Data}
\usage{
LDA.ms(
  ps = ps.cs,
  group = "Group",
  Top = 100,
  p.lvl = 0.05,
  lda.lvl = 2,
  seed = 11,
  adjust.p = F
)
}
\arguments{
\item{ps}{A `phyloseq` object containing metabolomic data, including the metabolite abundance table and sample metadata.}

\item{group}{A character string specifying the column name in the sample metadata that contains group information (e.g., treatment groups). Default is `"Group"`.}

\item{Top}{An integer specifying the number of top metabolites to retain for analysis based on their abundance. Default is `100`.}

\item{p.lvl}{A numeric value specifying the p-value cutoff for significance in Kruskal-Wallis tests. Default is `0.05`.}

\item{lda.lvl}{A numeric value specifying the minimum LDA score threshold for significance. Default is `2`.}

\item{seed}{An integer used to set the random seed for reproducibility of results. Default is `11`.}

\item{adjust.p}{A logical value indicating whether to adjust p-values using the false discovery rate (FDR) method. Default is `FALSE`.}
}
\value{
A list containing:
\itemize{
  \item `lefse_lists`: A data frame containing metabolite IDs, colors (for visualization), and group information for significant metabolites.
  \item `taxtree`: A data frame with detailed information about significant metabolites, including LDA scores, p-values, FDR values, and classification results.
}
}
\description{
The `LDA.ms` function performs Linear Discriminant Analysis (LDA) on metabolomic data in a `phyloseq` object. It identifies significant metabolites based on Kruskal-Wallis tests and LDA scores, providing a list of differentially abundant metabolites and their contributions across groups.
}
\details{
The function is designed to analyze metabolomic data by identifying significant metabolites associated with different experimental groups. The main steps include:
\enumerate{
  \item Filtering metabolites with zero counts across all samples and retaining the top metabolites based on their abundance.
  \item Performing Kruskal-Wallis tests for each metabolite to compute raw and adjusted p-values.
  \item Conducting LDA using the `MASS::lda` function to compute the discriminative scores for each metabolite.
  \item Identifying the maximum contributing group for each metabolite based on LDA scores.
  \item Filtering significant metabolites based on user-defined thresholds for p-values and LDA scores.
  \item Assigning colors to significant metabolites for downstream visualization (e.g., in graph-based plots).
}
}
\examples{
\dontrun{
tablda = LDA.ms(ps = ps.ms,Top = 100,p.lvl = 0.05,lda.lvl = 1,seed = 11,adjust.p = F)
p35 <- lefse_bar(taxtree = tablda[[2]])
dat = tablda[[2]]
}

}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
