% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefse.trans.R
\name{lefse.trans}
\alias{lefse.trans}
\title{Perform differential analysis using Lefse}
\usage{
lefse.trans(ps = ps, group = "Group", alpha = 0.05)
}
\arguments{
\item{ps}{A phyloseq format file used as an alternative for the input containing transcriptome functional composition table, tax, and sample metadata.}

\item{group}{Column name for groupID in map table(sample metadata).}

\item{alpha}{Significance level, default is 0.05.}
}
\value{
A list containing the results of significant differences in genes abundance between different groups.
}
\description{
This function is used to analyze transcriptome functional composition
 data  by conducting Lefse analysis to detect significant differences in genes
 abundance between different groups.
}
\examples{
\dontrun{
data(ps.trans)
ps = ps.trans\%>\% filter_taxa(function(x) sum(x ) > 5 , TRUE)
dat = lefse.trans(ps = ps\%>\% filter_OTU_ps(50),group =  "Group",alpha = 0.05)
dat1 = dat$WT_OE
head(dat1)
dat2 = dat$WT_KO
head(dat2)
dat3 = dat$OE_KO
head(dat3)
}
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
