% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefse.bar.R
\name{lefse_bar2}
\alias{lefse_bar2}
\title{Visualization of characteristic microorganisms only at OTU/ASV level}
\usage{
lefse_bar2(ps = pst, taxtree = tablda[[2]])
}
\arguments{
\item{ps}{phyloseq object of microbiome,including the Classified information of Characteristic microorganisms.}

\item{taxtree}{LDA.micro funtion output or a data frame containing taxonomic information and LDA results.}
}
\value{
A ggplot object representing the bar plot of the Lefse analysis,only OTU/ASV level.
}
\description{
This function generates a customized bar plot for Lefse analysis results, displaying the LDA scores
of only asv/otu classification levels. It arranges the bars based on the LDA scores
and colors them according to the Group.
}
\examples{
tablda = LDA.micro(ps = ps.16s,
Top = 100,
p.lvl = 0.05,
lda.lvl = 1,
seed = 11,
adjust.p = F)
p <- lefse_bar2(ps = ps.16s,taxtree = tablda[[2]])
p
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
