% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadingPCA.metf.R
\name{loadingPCA.metf}
\alias{loadingPCA.metf}
\title{PCA loading Matrix screening of characteristic genes}
\usage{
loadingPCA.metf(ps = ps, Top = 20)
}
\arguments{
\item{ps}{A phyloseq format file used as an alternative for the input containing metagenome functional composition table,
metagenome functional classification table, and sample metadata.}

\item{Top}{The top genes to visualize.}
}
\value{
A list object containing the following components:
\item{p}{A PCA correlation plot of the selected number of characteristic genes and
 the correlation decreases from top to bottom.}
\item{index}{Data frame containing the PCA load matrix and relative abundance of all genes.}
}
\description{
This function conducts PCA analysis on metagenome functional composition data to
extract the loading Matrix, screen for characteristic genes and visualize them.
The importance of the variables is  sorted according to
the square value of the correlation between the variable and the PC1 axis.
}
\examples{
ps =ps.kegg \%>\% filter_OTU_ps(Top = 1000)
res = loadingPCA.metf(ps = ps,Top = 20)
p = res[[1]]
p
dat = res[[2]]
dat
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
