% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadingPCA.omics.R
\name{loadingPCA.omics}
\alias{loadingPCA.omics}
\title{PCA loading Matrix screening of characteristic multi-omics data}
\usage{
loadingPCA.omics(ps = ps, Top = 20)
}
\arguments{
\item{ps}{A phyloseq format file used as an alternative for the input containing otu, tax, and map.}

\item{Top}{The top microorganisms to visualize.}
}
\value{
A list object containing the following components:
\item{p}{A PCA correlation plot of the selected number of characteristic microorganisms and
 the correlation decreases from top to bottom.}
\item{index}{Data frame containing the PCA load matrix and relative abundance of all microorganisms.}
}
\description{
This function conducts PCA analysis on multi-omics data to
extract the loading Matrix, screen for characteristic and visualize them.
The importance of the variables is  sorted according to
the square value of the correlation between the variable and the PC1 axis.
}
\examples{
res = loadingPCA.omics(ps = ps03,Top = 20)
p34.1 = res[[1]]
p34.1
dat = res[[2]]
dat
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
