% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantal.micro.R
\name{mantal.micro}
\alias{mantal.micro}
\title{Mantel Test and Visualization for Microbial Communities}
\usage{
mantal.micro(ps = ps, method = "spearman", group = "Group", ncol = 5, nrow = 2)
}
\arguments{
\item{ps}{A `phyloseq` object containing the microbial community data and metadata.}

\item{method}{A character string specifying the correlation method to be used in the Mantel test.
Options include `"pearson"`, `"spearman"` (default), and `"kendall"`.}

\item{group}{A character string indicating the name of the grouping variable in the metadata.
This variable will be used to split and analyze the data.}

\item{ncol}{An integer specifying the number of columns in the plot layout. Default is 5.}

\item{nrow}{An integer specifying the number of rows in the plot layout. Default is 2.}
}
\value{
This function returns a list containing the following:
- Mantel test results for each group.
- A visualization of the Mantel test results grouped by the specified factor.
}
\description{
The `mantal.micro` function performs a Mantel test to evaluate the correlation between two distance matrices
(e.g., microbial community dissimilarities and environmental distances) based on a specified correlation method.
It also groups the data by a specified factor and provides a visualization of the results.
The function allows customization of the layout of the plots, such as the number of columns and rows.
}
\examples{
# Example usage:
result <- mantal.micro(ps = ps.16s,method =  "spearman",group = "Group",ncol = gnum,nrow = 1)
data <- result[[1]]
data
p3_7 <- result[[2]]
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
