% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maptree.micro.R
\name{maptree.micro}
\alias{maptree.micro}
\title{Generate a Taxonomic Tree Map for Microbial Composition}
\usage{
maptree.micro(ps = ps, Top = 200, labtab = NULL, seed = 1)
}
\arguments{
\item{ps}{A `phyloseq` object containing microbiome data.}

\item{Top}{An integer specifying the number of top OTUs to include based on abundance. Default is `200`.}

\item{labtab}{Optional. A table for labeling the nodes in the taxonomic tree. If `NULL`, default labels are used.}

\item{seed}{An integer for setting the random seed for reproducibility. Default is `1`.}
}
\value{
A list containing:
\describe{
  \item{Plot}{A plot object representing the taxonomic tree map with abundance values.}
  \item{Data}{A data frame containing taxonomic information and abundance values used for plotting.}
}
}
\description{
This function constructs a taxonomic tree map from a `phyloseq` object, visualizing microbial taxonomy at various levels (Kingdom to OTU).
It uses hierarchical relationships between taxonomic levels to generate a tree structure and calculate abundance values for visualization.
}
\details{
The function performs the following steps:
\itemize{
  \item Filters the input `phyloseq` object to retain the top `Top` OTUs based on abundance.
  \item Constructs hierarchical taxonomic labels for each OTU, including all levels from Kingdom to Species.
  \item Builds a tree structure using parent-child relationships between taxonomic levels.
  \item Calculates mean abundance values for each OTU.
  \item Merges taxonomic and abundance data to prepare a data frame for plotting.
  \item Generates a taxonomic tree map using either default or custom labels (`labtab`).
}

This function is particularly useful for exploring microbial community structures and visualizing hierarchical taxonomic relationships.
}
\examples{
\dontrun{
tax = ps.16s \%>\% vegan_tax() \%>\% as.data.frame()
ps.bac <- ps.16s \%>\% subset_taxa.wt("Kingdom", "Bacteria")
res = maptree.micro(ps = ps.bac, Top = 100, labtab = NULL, seed = 11)
p20 = res[["Plot"]]
dat = res[["Data"]]
}

}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
