% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann.kegg.R
\name{match_KEGG}
\alias{match_KEGG}
\title{Match KEGG metabolites based on input metabolite(s) ID, KEGG database}
\usage{
match_KEGG(match, target_column, target_matrix)
}
\arguments{
\item{match}{The metabolite(s) to be matched.}

\item{target_column}{The target column for matching in the metabolite database.}

\item{target_matrix}{The matrix containing  the metabolite database.}
}
\value{
A list of matched KEGG metabolites based on the input metabolite(s) ID, KEGG database.
}
\description{
The match_KEGG function performs custom matching operations for retrieving
KEGG metabolite information based on the input match criteria, target column, and target matrix.
}
\examples{
library(dplyr)
tax= ps.ms \%>\% vegan_tax() \%>\%as.data.frame()
id = tax$Metabolite
tax2 = ann.kegg2(id)
head(tax2)
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
