% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Micro_tern.metf.R
\name{Micro_tern.metf}
\alias{Micro_tern.metf}
\title{Generate Ternary Plots for Metagenome functional composition data}
\usage{
Micro_tern.metf(ps = ps, color = "Level1")
}
\arguments{
\item{ps}{A `phyloseq` object containing metagenome functional composition table.}
}
\value{
A list containing the following elements:
\describe{
  \item{plot}{A list of ternary plots (`ggtern` objects) for different group combinations.}
  \item{dataplot}{A data frame with combined genes abundance and taxonomic information.}
  \item{groups}{A matrix showing all possible combinations of three groups used for ternary plots.}
}
}
\description{
This function generates ternary plots for visualizing metagenome functional composition data in three groups.
It uses relative abundance data from a `phyloseq` object and displays metagenome functional composition distribution across specified groupings.
}
\details{
This function performs the following steps:
\itemize{
  \item Standardizes the metagenome functional composition table to relative abundances using `phyloseq::transform_sample_counts`.
  \item Splits the metagenome functional composition data into groups based on the metadata column `Group`.
  \item Computes the mean abundance for each gene within each group.
  \item Generates ternary plots for all possible combinations of three groups using the `ggtern` package.
}

The plots visualize the relative abundances of genes in three groups, with points sized by their mean abundance and colored by their taxonomic classification (pathway).
}
\examples{
\dontrun{
ps =ps.kegg \%>\% filter_OTU_ps(Top = 1000)
ps1 = ps \%>\% filter_OTU_ps(500)
res = Micro_tern.metf(ps1)
p15 = res[[1]]
dat =  res[[2]]
}

}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
