% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Micro_tern.micro.R
\name{Micro_tern.micro}
\alias{Micro_tern.micro}
\title{Generate Ternary Plots for Microbial Data}
\usage{
Micro_tern.micro(ps = ps)
}
\arguments{
\item{ps}{A `phyloseq` object containing microbiome data.}
}
\value{
A list containing the following elements:
\describe{
  \item{plot}{A list of ternary plots (`ggtern` objects) for different group combinations.}
  \item{dataplot}{A data frame with combined OTU abundance and taxonomic information.}
  \item{groups}{A matrix showing all possible combinations of three groups used for ternary plots.}
}
}
\description{
This function generates ternary plots for visualizing microbial community data in three groups. It uses relative abundance data from a `phyloseq` object and displays OTU distribution across specified groupings.
}
\details{
This function performs the following steps:
\itemize{
  \item Standardizes the OTU table to relative abundances using `phyloseq::transform_sample_counts`.
  \item Splits the OTU data into groups based on the metadata column `Group`.
  \item Computes the mean abundance for each OTU within each group.
  \item Generates ternary plots for all possible combinations of three groups using the `ggtern` package.
}

The plots visualize the relative abundances of OTUs in three groups, with points sized by their mean abundance and colored by their taxonomic classification (Phylum).
}
\examples{
\dontrun{
ps1 = ps.16s \%>\% filter_OTU_ps(500)
res = Micro_tern.micro(ps1)
p15 = res[[1]]
dat =  res[[2]]
}

}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
