% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkegg_function.R
\name{mkegg_function}
\alias{mkegg_function}
\title{Reaction-Based KEGG Pathway Analysis for Phyloseq Object}
\usage{
mkegg_function(ps = ps)
}
\arguments{
\item{ps}{A `phyloseq` object containing OTU data to be transformed into KEGG reactions.}

\item{dat}{A data frame containing KO identifiers and associated reactions. If `NULL`, defaults to `EasyMultiOmics.db::db.ko.reaction`.}

\item{dat2}{A data frame containing reactions and descriptions. If `NULL`, defaults to `EasyMultiOmics.db::db.reaction`.}
}
\value{
A `phyloseq` object with summarized OTU data and KEGG reaction information.
}
\description{
This function processes an OTU table from a `phyloseq` object, maps KO identifiers to KEGG reactions,
and creates a new `phyloseq` object that associates reactions with the OTU data.
}
\details{
This function performs several steps:
\itemize{
  \item Loads KO-reaction data from `EasyMultiOmics.db::db.ko.reaction` by default or the provided `dat`.
  \item Transforms the OTU table in `ps` to KO identifiers using `vegan_otu()`.
  \item Joins KO identifiers with KEGG reaction data.
  \item Summarizes the data by reaction and creates a new `phyloseq` object with summarized OTU data.
  \item Adds reaction descriptions from `EasyMultiOmics.db::db.reaction` or the provided `dat2`.
}
}
\examples{
# Assuming you have a phyloseq object `ps`
reaction_ps <- reaction_function(ps)

}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
