% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_associated_metabolites.R
\name{ml_feature_selection}
\alias{ml_feature_selection}
\title{Machine Learning Feature Selection}
\usage{
ml_feature_selection(microbe_abundance, metabolome_data)
}
\arguments{
\item{microbe_abundance}{Numeric vector of microbe abundance}

\item{metabolome_data}{Matrix of metabolome data}
}
\value{
List of ML results
}
\description{
Uses multiple ML approaches to identify important metabolites
}
