% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_micro.rf.omics.R
\name{ms_micro.rf.omics}
\alias{ms_micro.rf.omics}
\title{Random Forest Analysis for Microbiome Data Using NMDS}
\usage{
ms_micro.rf.omics(ps = ps, env = env, seed = 1)
}
\arguments{
\item{ps}{A phyloseq object containing the microbiome data. This object should include OTU data and sample metadata.}

\item{env}{A data frame or vector containing the environmental variables to be used in the analysis.}

\item{seed}{An integer for setting the random seed to ensure reproducibility (default is 1).}
}
\value{
A list containing two elements:
\item{df}{A data frame containing the importance of each feature (environmental variable) based on Random Forest analysis.}
\item{p}{A ggplot object representing a bar plot of feature importance.}
}
\description{
The function first converts the OTU (Operational Taxonomic Unit) table from a phyloseq object into a
numerical matrix, computes a Bray-Curtis distance matrix, performs NMDS ordination, and then applies
Random Forest to determine the importance of environmental variables on the first NMDS axis.
A bar plot is generated to visualize the feature importance based on Random Forest results.
}
\examples{
\dontrun{
ps.tem = ps.micro \%>\% filter_OTU_ps(500)
res = loadingPCA.ms(ps = ps.ms)
dat = res[[2]]
dat$id = row.names(dat)
id = dat \%>\% arrange(desc(PCone)) \%>\% head(15) \%>\% .$id
ftab = ps.ms \%>\% scale_micro() \%>\% subset_taxa(row.names(tax_table(ps.ms)) \%in\% c(id) ) \%>\% vegan_otu() \%>\%as.data.frame()
result <- ms_micro.rf.omics(ps = ps, env = env, seed = 42)
}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
