% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuiPlot_FEAST.R
\name{MuiPlot_FEAST}
\alias{MuiPlot_FEAST}
\title{Multiple FEAST Source Contribution Visualizations}
\usage{
MuiPlot_FEAST(data = result)
}
\arguments{
\item{data}{A matrix or data frame where rows represent sources, and columns represent samples or groups. Typically, this is the output from FEAST analysis. Default is `result`.}
}
\value{
A combined ggplot2 object with polar bar charts for each sample/group. Each chart includes source contributions and their respective percentages.
}
\description{
The `MuiPlot_FEAST` function creates multiple polar bar charts (pie charts) to visualize the source contributions for each sample or group in the result matrix from FEAST analysis. Each chart represents the relative contributions of sources to a particular sample/group.
}
\details{
The function iterates through the columns of the input matrix, calculates the mean contribution of each source for a given sample/group, normalizes the contributions, and visualizes them as polar bar charts. The charts are combined into a single plot with a shared legend for easy comparison.
}
\examples{
\dontrun{
result = FEAST.micro(ps = ps.16s,group = "Group",sinkG = "OE",sourceG = c("WT","KO"))
p2 = MuiPlot_FEAST(data = result)
}

}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
