% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naivebayes.ms.R
\name{naivebayes.ms}
\alias{naivebayes.ms}
\title{naiveBayes model screening of characteristic metabolites}
\usage{
naivebayes.ms(ps = ps, top = 20, seed = 1010, k = 5)
}
\arguments{
\item{ps}{A phyloseq format file used as an alternative for the input containing metabolite composition table,
metabolite classification table, and sample metadata.}

\item{top}{The top metabolites to consider.}

\item{seed}{The random seed for reproducibility.}

\item{k}{The number of folds for cross-validation.}
}
\value{
A list object including the following components:
\item{Accuracy}{The average accuracy of the naiveBayes model.}
\item{Importance}{A data frame showing the feature importance ranked in descending order.}
}
\description{
naiveBayes, one of the machine learning methods, was used to screen for characteristic
metabolites, and the model was evaluated using k-fold cross-validation.
}
\examples{
library(dplyr)
library(ggClusterNet)
library(caret)
res = naiveBayes.micro(ps=ps.ms, top = 20, seed = 1010, k = 5)
accuracy = res[[1]]
accuracy
importance = res[[2]]
importance
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
