% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.sci.R
\name{nature_style}
\alias{nature_style}
\title{Add Nature-style plot modifications}
\usage{
nature_style(plot, expand_x = c(0.02, 0), expand_y = c(0.02, 0))
}
\arguments{
\item{plot}{A ggplot object}

\item{expand_x}{X-axis expansion (default: c(0.02, 0))}

\item{expand_y}{Y-axis expansion (default: c(0.02, 0))}
}
\value{
Modified ggplot object
}
\description{
Add Nature-style plot modifications
}
\examples{
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
nature_style(p)
}
