% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.pip.metf.R
\name{network.pip.metf}
\alias{network.pip.metf}
\title{Metagenome functional composition related network}
\usage{
network.pip.metf(
  otu = NULL,
  tax = NULL,
  map = NULL,
  ps = NULL,
  N = 0,
  ra = NULL,
  order,
  big = TRUE,
  select_layout = FALSE,
  layout_net = "model_maptree2",
  r.threshold = 0.6,
  p.threshold = 0.05,
  maxnode = 2,
  method = "spearman",
  label = FALSE,
  lab = "elements",
  group = "Group",
  fill = "Phylum",
  size = "igraph.degree",
  zipi = FALSE,
  clu_method = "cluster_fast_greedy",
  step = 100,
  ram.net = FALSE,
  R,
  ncpus
)
}
\arguments{
\item{ps}{A phyloseq format file used as an alternative for the input containing metagenome functional composition table, tax, and sample metadata.}

\item{N}{filter metagenome functional composition tables by abundance.The defult, N=0, extract the top N number relative abundance of gene. e.g 100}

\item{big}{TRUE or FALSE the number of micro data was so many (> 300),you can chose TRUE}

\item{select_layout}{TURE or FALSE}

\item{layout_net}{select layout from ggClusterNet}

\item{r.threshold}{The defult, r.threshold=0.6, it represents the correlation that the absolute value
of the correlation threshold is greater than 0.6. the value range of correlation threshold from 0 to 1.}

\item{p.threshold}{The defult, p.threshold=0.05, it represents significance threshold below 0.05.}

\item{method}{method for Correlation calculation,method="pearson" is the default value. The alternatives to be passed to cor are "spearman" and "kendall".}

\item{label}{Whether to add node label.}

\item{group}{Separate Group.}

\item{fill}{fill coulor of node}

\item{size}{node size}

\item{zipi}{zipi Calculation}

\item{step}{Random network sampling times}

\item{R}{repeat number of p value calculate}

\item{ncpus}{number of cpus used for sparcc}

\item{lay}{layout which network show}

\item{path}{save path of all of network analyse.}
}
\value{
list which contains gene correlation matrix and ggplot2 plot.
}
\description{
This function is designed to generate graphical representations related to network analysis and compute various network properties.
}
\examples{
library(ggClusterNet)
library(igraph)
data(ps.kegg)
result = network.pip.metf(N = 400,big = TRUE,select_layout = FALSE,layout_net = "model_maptree2",
r.threshold = 0.6,p.threshold = 0.05,maxnode = 2,label = FALSE,lab = "elements",group = "Group",
size = "igraph.degree",zipi = TRUE,ram.net = TRUE,clu_method = "cluster_fast_greedy",step = 100,
R=10,ncpus = 1,fill = "Pathway")
dat = result[[2]]
cortab = dat$net.cor.matrix$cortab
plot = result[[1]]
p0 = plot[[1]]
p0
}
\author{
Contact: Tao Wen \email{taowen@njau.edu.cn};Penghao Xie \email{2019103106@njau.edu.cn}.
}
