% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnet.metf.R
\name{nnet.metf}
\alias{nnet.metf}
\title{Neural network screening of characteristic genes}
\usage{
nnet.metf(ps = ps, top = 20, seed = 1010, k = 5)
}
\arguments{
\item{ps}{A phyloseq format file used as an alternative for the input containing metagenome functional composition table,
metagenome functional classification table, and sample metadata.}

\item{top}{The top genes to consider.}

\item{seed}{The random seed for reproducibility.}

\item{k}{The number of folds for cross-validation.}
}
\value{
A list object including the following components:
\item{Accuracy}{The average accuracy of the neural network model.}
\item{Importance}{A data frame showing the feature importance ranked in descending order.}
}
\description{
A neural network, one of the machine learning methods, was used to screen for characteristic
genes, and the model was evaluated using k-fold cross-validation.
}
\examples{
library(dplyr)
library(ggClusterNet)
ps =ps.kegg \%>\% filter_OTU_ps(Top = 1000)
res =nnet.metf(ps=ps, top = 100, seed = 1010, k = 5)
accuracy = res[[1]]
accuracy
importance = res[[2]]
importance
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
