% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.ms.R
\name{normalize.ms}
\alias{normalize.ms}
\title{Normalize metabolite data using specified method}
\usage{
normalize.ms(ps, method = "rela")
}
\arguments{
\item{ps}{A phyloseq format file used as an alternative for the input containing metabolite, tax, and sample data.}

\item{method}{The normalization method to be applied,default is "rela".}
}
\value{
The new phyloseq format file by Selected method after the standardization.
}
\description{
The normalize.ms function normalizes  metabolite data based on the specified method.
Three normalization methods are supported: "rela" for relative abundance normalization,
"sampling" for sampling normalization (adjusting counts to make the total count of each sample equal to the average total count),
and "log" for log transformation.
}
\examples{
ps.ms2 = normalize.ms(ps = ps.ms,method = "rela")
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
