% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oplsda.ms.R
\name{oplsda.ms}
\alias{oplsda.ms}
\title{Perform Orthogonal Partial least squares Discriminant Analysis (OPLS-DA) on metabolite data}
\usage{
oplsda.ms(
  ps = ps,
  Group = "Group",
  orthoI_num = 3,
  oplspath = NULL,
  vipplot_top = 30,
  ncol = ncol,
  nrow = nrow
)
}
\arguments{
\item{ps}{A phyloseq format file used as an alternative for the input containing metabolite composition table,
metabolite classification table, and sample metadata.}

\item{Group}{Column name for groupID in map table(sample metadata).}

\item{orthoI_num}{The number of orthogonal components in the OPLS-DA.}

\item{oplspath}{Path to save the output files.}

\item{vipplot_top}{Top number of VIP features to consider in plots.}

\item{ncol}{Number of columns for arranging plots.}

\item{nrow}{Number of rows for arranging plots.}
}
\description{
The oplsda.ms function conducts Orthogonal Partial least squares Discriminant Analysis (OPLS-DA) on metabolite data using the ropls package.
It visualizes the results of OPLS-DA including score plots, vip plots, and Permutation plots.
}
\examples{
{library(pacman)
library(ggsci)
library(ropls)
res = oplsda.ms(ps = ps.ms, ncol=3,nrow = 1)
p1 = res[[1]]
p1
p2 = res[[2]]
p2
p3 = res[[3]]
p3
dat1 = res[[4]]
dat1
dat2 = res[[4]]
dat2
dat3 = res[[6]]
dat3}
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
