% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinate.micro.R
\name{ordinate.micro}
\alias{ordinate.micro}
\title{Beta diversity plotting}
\usage{
ordinate.micro(
  otu = NULL,
  tax = NULL,
  map = NULL,
  ps = NULL,
  group = "Group",
  dist = "bray",
  method = "PCoA",
  Micromet = "adonis",
  pvalue.cutoff = 0.05
)
}
\arguments{
\item{otu}{OTU/ASV table;}

\item{tax}{taxonomy table;}

\item{map}{Sample metadata;}

\item{ps}{A phyloseq format file used as an alternative for the input containing otu, tax, and map.}

\item{group}{group ID;}

\item{dist}{distance type, including "unifrac" "wunifrac" "dpcoa" "jsd" "manhattan" "euclidean"   "canberra" "bray" "kulczynski"  "jaccard" "gower" "altGower" "morisita" "horn" "mountford"  "raup" "binomial"  "chao"  "cao" "w"  "-1"  "c" "wb"  "r"   "I"  "e" "t" "me"   "j"  "sor"  "m"   "-2"  "co";}

\item{method}{DCA, CCA, RDA, NMDS, MDS, PCoA, PCA, LDA;}

\item{Micromet}{statistics by adonis/anosim/MRPP;}

\item{pvalue.cutoff}{Pvalue threshold;}
}
\value{
list object including plot, stat table
}
\description{
Input otutab, metadata and tree or phyloseq object; support 47 distance type (bray, unifrac, wunifrac ...),  8 ordination method (PCoA, NMDS, ...); output ggplot2 figure, data and statistical test result.
}
\details{
By default, input phyloseq object include metadata, otutab and tree
The available diversity indices include the following:
\itemize{
\item{most used indices: bray, unifrac, wunifrac}
\item{other used indices: manhattan, euclidean, jaccard ...}
}
}
\examples{
result = ordinate.micro(ps = ps.16s, group = "Group", dist = "bray",method = "PCoA", Micromet = "anosim",
pvalue.cutoff = 0.05)
p3_1 = result[[1]]
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
