% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinate.ms.R
\name{ordinate.ms}
\alias{ordinate.ms}
\title{Ordination analysis and plotting}
\usage{
ordinate.ms(
  otu = NULL,
  tax = NULL,
  map = NULL,
  ps = NULL,
  group = "Group",
  dist = "bray",
  method = "PCoA",
  Micromet = "adonis",
  pvalue.cutoff = 0.05,
  pair = TRUE
)
}
\arguments{
\item{otu}{Metabolite composition table.}

\item{tax}{Metabolite classification table.}

\item{map}{Sample metadata.}

\item{ps}{A phyloseq format file used as an alternative for the input containing metabolite, tax, and sample metadata.}

\item{group}{Column name for groupID in map table(sample metadata).}

\item{dist}{Distance type, including "unifrac","wunifrac","manhattan","euclidean","bray","jaccard"... .}

\item{method}{DCA, CCA, RDA, NMDS, MDS, PCoA, PCA, LDA, t-sne.}

\item{Micromet}{statistics by adonis/anosim/MRPP;}

\item{pvalue.cutoff}{Pvalue threshold.}

\item{pair}{A logical value for whether to perform analysis between any two groups.}
}
\value{
A list containing the following components:
\item{p2}{Standard plot of the ordination results.}
\item{points}{Data frame with coordinates of samples in the ordination space.}
\item{p3}{Annotated plot with sample labels.}
\item{pairResult}{Results of pairwise tests if applicable.}
\item{title1}{Overall inspection results by anosim or adonis.}
\item{eig}{Eigenvalues or variance explained by each axis.}
}
\description{
Input metabolite table, metadata and tree or phyloseq object;
support 47 distance type (bray, unifrac, wunifrac ...),
8 ordination method (PCoA, NMDS, ...);
output ggplot2 figure, data and statistical test result.
}
\examples{
result = ordinate.ms(ps = ps.ms, group = "Group", dist = "bray",
method = "PCoA", Micromet = "anosim", pvalue.cutoff = 0.05,pair = F)
p3_1 = result[[1]]
p3_1
p3_2 = result[[3]]
p3_2
plotdata =result[[2]]
head(plotdata)
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
