% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_base.micro.R
\name{p_base.micro}
\alias{p_base.micro}
\title{Visualize taxonomic information using circular tree layout}
\usage{
p_base.micro(ps, Top = 300)
}
\arguments{
\item{ps}{A phyloseq format file used as an alternative for the input containing otu, tax, and map.}

\item{Top}{The top microorganisms to convert and visualize.}
}
\value{
A ggplot object displaying taxonomic information in a circular tree layout.
}
\description{
convert dataframe contained hierarchical relationship or other classes to
treedata class and visualize the circular tree.
}
\examples{
p1 <- p_base.micro(ps.16s,Top = 100)
p1
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
