% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.amp.R
\name{package.amp}
\alias{package.amp}
\title{Load Required R Packages for Data Analysis}
\usage{
package.amp()
}
\value{
This function does not return any value. It is used for loading required libraries.
}
\description{
The function loads various R libraries commonly used for microbiome analysis, data manipulation,
and visualization. It includes packages like `phyloseq`, `ggClusterNet`, `EasyStat`, and more.
This function is intended to streamline the initial setup for data analysis by loading multiple
libraries in one call, saving time and avoiding repetitive loading of the same libraries in
separate scripts.
}
\examples{
\dontrun{
# Load all required packages for microbiome analysis
package.amp()
}
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njau.edu.cn}
}
