% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairMetaTest.metf.R
\name{pairMetaTest.metf}
\alias{pairMetaTest.metf}
\title{Beta diversity statistics by adonis/anosim/MRPP in pair}
\usage{
pairMetaTest.metf(ps = ps, Micromet = "adonis", dist = "bray")
}
\arguments{
\item{ps}{A phyloseq format file used as an alternative for the input containing metagenome functional composition table, tax, and sample metadata.}

\item{Micromet}{Statistics default by adonis, alternative anosim or MRPP;}

\item{dist}{Distance type, including "unifrac" "wunifrac" "dpcoa" "jsd" "manhattan" "euclidean"   "canberra" "bray" "kulczynski"  "jaccard" "gower" "altGower" "morisita" "horn" "mountford"  "raup" "binomial"  "chao"  "cao" "w"  "-1"  "c" "wb"  "r"   "I"  "e" "t" "me"   "j"  "sor"  "m"   "-2"  "co";}

\item{otu}{Metagenome functional composition table.}

\item{map}{Sample metadata;}

\item{tree}{Tree/nwk file;}

\item{group}{Column name for groupID in map table(sample metadata).}

\item{method}{DCA, CCA, RDA, NMDS, MDS, PCoA, PCA, LDA;}

\item{pvalue.cutoff}{Pvalue threshold, default in 0.05.}
}
\value{
stat table
}
\description{
Input phyloseq object, test method and distance type.
}
\details{
By default, input phyloseq object include metadata and metagenome functional composition table
The available diversity indices include the following:
\itemize{
\item{most used indices: bray unifrac wunifrac}
\item{other used indices: dpcoa jsd manhattan euclidean canberra kulczynski jaccard gower altGower morisita horn mountford raup binomial chao cao w -1 c wb r I e t me j sor m -2 co}
}
}
\examples{
ps =ps.kegg \%>\% filter_OTU_ps(Top = 1000)
dat2 = pairMicroTest.metf(ps = ps, Micromet = "MRPP", dist = "bray")
dat2
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
