% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairMicroTest.metm2.R
\name{pairMicroTest.metm2}
\alias{pairMicroTest.metm2}
\title{Pairwise Microbial Community Statistical Testing}
\usage{
pairMicroTest.metm2(ps, Micromet = "adonis", dist = "bray")
}
\arguments{
\item{ps}{A `phyloseq` object containing OTU/ASV data and sample metadata.}

\item{Micromet}{Statistical method: `"adonis"` (default), `"MRPP"`, `"anosim"`.}

\item{dist}{Distance metric for community dissimilarity (default = `"bray"`).}
}
\value{
A data.frame with pairwise group comparisons, statistic, and p-value.
}
\description{
Perform pairwise statistical testing on microbial community data
(Adonis, MRPP, or ANOSIM) using a phyloseq object and a given distance metric.
}
\examples{
\dontrun{
pairMicroTest.metm2(ps, Micromet = "permdisp", dist = "bray")
pairMicroTest.metm2(ps, Micromet = "mantel", dist = "bray")
}
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njau.edu.cn}
}
