% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairordinateTest.omics.R
\name{pairordinateTest.omics}
\alias{pairordinateTest.omics}
\title{PairwiseCommunity Statistical Testing}
\usage{
pairordinateTest.omics(ps = ps, Micromet = "adonis", dist = "bray")
}
\arguments{
\item{ps}{A `phyloseq` object containing microbial community data, including the OTU table and sample metadata.}

\item{Micromet}{A string specifying the statistical method to use. Options are `"adonis"` (default), `"MRPP"`, or `"anosim"`.}

\item{dist}{A string specifying the distance metric to use. Default is `"bray"` (Bray-Curtis distance).}
}
\value{
stat table
}
\description{
The `pairMicroTest.omics` function performs pairwise statistical testing on microbial community data using selected metrics (Adonis, MRPP, or ANOSIM). It calculates statistical differences between groups in a `phyloseq` object based on a specified distance metric.
}
\examples{
dat2 = pairMicroTest.omics(ps = ps.03, Micromet = "MRPP", dist = "bray")
dat2
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
