% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrich.ms.R
\name{pathway_enrich.ms}
\alias{pathway_enrich.ms}
\title{Perform KEGG pathway enrichment analysis for metabolite data}
\usage{
pathway_enrich.ms(ps = ps, dif.method = "wilcox")
}
\arguments{
\item{ps}{A phyloseq format file used as an alternative for the input containing metabolite composition table, metabolite classification table, and sample metadata.}

\item{dif.method}{Method for differential analysis, default is "wilcox".}
}
\value{
A list containing the following components:
\describe{
  \item{Plots}{List of ggplot objects representing the pathway enrichment analysis plots.}
  \item{Plotdata}{List of data frames containing pathway enrichment analysis results.}
}
}
\description{
Through the difference test, the significantly enriched KEGG functional pathway was identified and visualized.
}
\examples{
\dontrun{
ps.ms3 = ps.ms \%>\% tax_glom_wt("KEGGID")
res2 = pathway_enrich.ms(ps = ps.ms3, dif.method = "wilcox")
res2$Plots$OE.WT.plot
res2$Plots$KO.OE.plot
}
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njau.edu.cn}
}
