% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrich.trans.R
\name{pathway_enrich.trans}
\alias{pathway_enrich.trans}
\title{KEGG Pathway Enrichment Analysis with Differential Expression}
\usage{
pathway_enrich.trans(ps, dif.method = "wilcox")
}
\arguments{
\item{ps}{A Phyloseq object containing OTU table and sample data. The `Group` column in the
sample data will be used to define the groups for the analysis.}

\item{dif.method}{A string specifying the differential analysis method to use (default is
"wilcox"). The method should correspond to the differential test used in the `statSuper()` function.}
}
\value{
A list containing:
\itemize{
  \item \code{plots}: A list of ggplot2 bar plots showing the enriched KEGG pathways for
  each group comparison.
  \item \code{plotdata}: A list of data frames containing the KEGG enrichment results for each
  comparison.
}
}
\description{
This function performs KEGG pathway enrichment analysis using differential expression results
from the `statSuper()` function. It identifies significantly enriched pathways based on the
given method (default is "wilcox") and generates bar plots of the top enriched pathways.
}
\examples{
\dontrun{
result <- pathway_enrich.trans(ps = ps.trans, dif.method = "wilcox")
result$plots  # View the plots of enriched KEGG pathways
}
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
