% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyRare_cpp.R
\name{phyRare}
\alias{phyRare}
\title{R-based Rarefaction for Phyloseq Objects}
\usage{
phyRare(ps, N = 3000, cores = 10)
}
\arguments{
\item{ps}{A \code{phyloseq} object.}

\item{N}{Integer. Rarefaction depth. Default 3000.}

\item{cores}{Integer. Number of cores for parallel computation. Default 10.}
}
\value{
A rarefied \code{phyloseq} object.
}
\description{
Performs rarefaction on a phyloseq object using R (vegan::rrarefy).
Supports parallel computation.
}
\examples{
ps_rarefied <- phyRare(ps, N = 3000, cores = 4)
}
