% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyRare_cpp.R
\name{phyRare_auto}
\alias{phyRare_auto}
\title{Automatic Rarefaction for Phyloseq Objects}
\usage{
phyRare_auto(ps, N = 3000, ...)
}
\arguments{
\item{ps}{A \code{phyloseq} object containing OTU/ASV table and sample data.}

\item{N}{Integer. Rarefaction depth (number of sequences per sample). Default is 3000.}

\item{...}{Additional parameters passed to \code{phyRare_cpp()} or \code{phyRare()}.}
}
\value{
A rarefied \code{phyloseq} object.
}
\description{
Automatically performs rarefaction on a phyloseq object.
It tries to use the faster \code{phyRare_cpp()} if available;
otherwise, it falls back to the R version \code{phyRare()}.
}
\examples{
ps_rarefied <- phyRare_auto(ps, N = 3000)
}
