% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewiseSEM.omics.R
\name{piecewiseSEM.omics}
\alias{piecewiseSEM.omics}
\title{Perform Piecewise Structural Equation Modeling (SEM) for Omics Data}
\usage{
piecewiseSEM.omics(
  ps.trans = ps.trans,
  ps.ms = ps.ms,
  ps.micro = ps.micro,
  ps.meta = ps.kegg,
  filter = 0.001
)
}
\value{
A list containing:
- `r_squared`: The R² value of the model.
- `aic`: The Akaike Information Criterion (AIC) value of the model.
- `bic`: The Bayesian Information Criterion (BIC) value of the model.
}
\description{
This function performs Piecewise Structural Equation Modeling (SEM) on multiple omics datasets (e.g., transcriptomics, metabolomics, microbiome, metagenome).
The function uses Principal Component Analysis (PCA) to extract the first principal component from each omics dataset, filters low abundance taxa,
and then builds a piecewise SEM model to examine the relationships between different omics layers.

The function takes multiple omics datasets (transcriptome, metabolome, microbiome, and metagenome), intersects their samples,
filters low abundance features, and performs PCA. Then, a piecewise SEM model is constructed to understand the relationships between
these datasets, such as the influence of transcriptomics on metabolomics, metabolomics on microbiomes, and the combined effects of
metabolomics and microbiomes on metagenomes. The function returns the R², AIC, and BIC values of the model.
}
\examples{
\dontrun{
library(vegan)
library(phyloseq)
library(dplyr)
library(ggClusterNet)
library(lavaan)
library(semPlot)
library(EasyMultiOmics)
data(ps.trans)
data(ps.ms)
data(ps.micro)
data(ps.kegg)
res <- piecewiseSEM.omics(ps.trans=ps.trans,ps.ms=ps.ms,ps.micro=ps.micro,
ps.meta=ps.kegg,filter=0.05)
res
}

}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njau.edu.cn}.
}
