% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_metabolite_selection.R
\name{plot.metabolite_selection_report}
\alias{plot.metabolite_selection_report}
\title{Plot Method for Selection Report}
\usage{
\method{plot}{metabolite_selection_report}(x, ...)
}
\arguments{
\item{x}{A metabolite_selection_report object}

\item{...}{Additional arguments (ignored)}
}
\description{
Plot Method for Selection Report
}
