% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_FEAST.R
\name{Plot_FEAST}
\alias{Plot_FEAST}
\title{FEAST Source Contribution Visualization}
\usage{
Plot_FEAST(data = result)
}
\arguments{
\item{data}{A matrix or data frame where rows represent sources, and columns represent contributions across samples. Typically, this is the output from FEAST analysis. Default is `result`.}
}
\value{
A ggplot2 object representing the polar bar chart with source contributions and their respective percentages.
}
\description{
The `Plot_FEAST` function generates a polar bar chart (pie chart) to visualize the relative contributions of different sources using FEAST (Fast Expectation-Maximization for Source Tracking).
}
\details{
The function calculates the mean contribution of each source across samples, normalizes the contributions to ensure they sum to 100%, and visualizes the contributions as a polar bar chart (pie chart). Each slice of the chart represents a source, and the label includes the source name and its percentage contribution.
}
\examples{
\dontrun{
result = FEAST.micro(ps = ps.16s,group = "Group",sinkG = "OE",sourceG = c("WT","KO"))
p <- Plot_FEAST(data = result)
}

}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
