% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_feature_selection.R
\name{plot_ms_results}
\alias{plot_ms_results}
\title{Visualize MS Feature Selection Results}
\usage{
plot_ms_results(results, top_n = 20)
}
\arguments{
\item{results}{ms_feature_selection 的返回}

\item{top_n}{展示前 N}
}
\description{
Visualize MS Feature Selection Results
}
