% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.ms.R
\name{plsda_ms}
\alias{plsda_ms}
\title{Perform Partial Least Squares Discriminant Analysis (PLS-DA) on metabolite data}
\usage{
plsda_ms(map = NULL, ps = NULL, count = NULL, Group = "Group")
}
\arguments{
\item{map}{A data frame containing sample metadata.}

\item{ps}{A phyloseq format file used as an alternative for the input containing metabolite composition table,
metabolite classification table, and sample metadata.}

\item{count}{A matrix of metabolite counts.}

\item{Group}{Column name for groupID in map table(sample metadata).}
}
\value{
A list containing the following components:
\item{p}{Plot showing the PLS-DA analysis results.}
\item{plotdata}{Data frame containing the coordinates of samples in the PLS-DA space.}
}
\description{
The plsda.ms function conducts Partial Least Squares Discriminant Analysis (PLS-DA) on metabolite data using the mixOmics package.
It visualizes the results of PLS-DA including score plots, explained variance, and ellipses.
}
\examples{
library(mixOmics)
library(ggplot2)
res = plsda_ms(ps=ps.ms,Group = "Group")
p11 = res [[1]]
p11
dat = res[[2]]
dat
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
