% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_associated_metabolites.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare Data for Analysis}
\usage{
prepare_data(phyloseq_obj, metabolome_mat, target_microbe_name)
}
\arguments{
\item{phyloseq_obj}{A phyloseq object}

\item{metabolome_mat}{Metabolome data matrix}

\item{target_microbe_name}{Name of target microbe}
}
\value{
A list containing aligned data
}
\description{
Extracts and aligns microbiome and metabolome data for analysis
}
