% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_rmt_scale_free2.R
\name{preprocess_phyloseq}
\alias{preprocess_phyloseq}
\title{Preprocess phyloseq data}
\usage{
preprocess_phyloseq(
  phyloseq_obj,
  min_abundance,
  min_prevalence,
  transformation
)
}
\description{
Preprocess phyloseq data
}
