% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_metabolite_selection.R
\name{print_selection_report}
\alias{print_selection_report}
\title{Print Selection Report}
\usage{
print_selection_report(report)
}
\arguments{
\item{report}{Selection report object}
}
\description{
Prints formatted selection report to console.
}
\keyword{internal}
