% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quare.line.omics.R
\name{quare.line.omics}
\alias{quare.line.omics}
\title{This function performs an analysis comparing two groups based on microbiome (16S rRNA) and metabolomics data.
It identifies differentially abundant operational taxonomic units (OTUs) and metabolites, then compares
their association through statistical tests. The function visualizes the comparisons using a scatter plot.}
\usage{
quare.line.omics(
  ps.16s = ps.16s,
  ps.ms = ps.ms,
  group1 = "micro",
  group2 = "ms",
  r.threshold = 0.6,
  p.threshold = 0.1,
  method = "spearman",
  top = 500
)
}
\arguments{
\item{ps.ms}{A data frame or tibble containing metabolomics data (default: `ps.ms`).}

\item{group1}{A string specifying the first group for comparison (default: `"micro"`).}

\item{group2}{A string specifying the second group for comparison (default: `"ms"`).}

\item{r.threshold}{A numeric value specifying the correlation threshold for filtering the microbiome data (default: 0.6).}

\item{p.threshold}{A numeric value specifying the p-value threshold for significance in statistical tests (default: 0.1).}

\item{method}{A string specifying the statistical test method (e.g., `"spearman"`) (default: `"spearman"`).}

\item{top}{An integer specifying the number of top OTUs or metabolites to consider for analysis (default: 500).}

\item{ps01}{A phyloseq object containing microbiome data (default: `ps.16s`).}
}
\value{
A list containing:
\item{dat}{A data frame with the merged OTU and metabolite data, with statistical results and category assignments.}
\item{plot}{A ggplot object visualizing the comparisons between the two groups.}
}
\description{
This function performs an analysis comparing two groups based on microbiome (16S rRNA) and metabolomics data.
It identifies differentially abundant operational taxonomic units (OTUs) and metabolites, then compares
their association through statistical tests. The function visualizes the comparisons using a scatter plot.
}
\examples{
results <- quare.line.omics(ps.16s =ps.16s ,
                            ps.ms = ps.ms,
                            group1 = "microbe",
                            group2 = "metabolite",
                            r.threshold = 0.7,
                            p.threshold = 0.05,
                            method = "spearman",
                            top = 100)

}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
