% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramdom_env.plot.R
\name{ramdom_env.plot}
\alias{ramdom_env.plot}
\title{Random Forest Analysis of Environmental Variables on NMDS Axis}
\usage{
ramdom_env.plot(ps = ps, env = env, seed = 1)
}
\arguments{
\item{ps}{A `phyloseq` object containing OTU/taxa abundance data and sample metadata.}

\item{env}{A data frame of environmental variables, where rows are samples and columns are variables.}

\item{seed}{An integer value to set the random seed for reproducibility. Default is `1`.}
}
\value{
A list containing:
\itemize{
  \item `df`: A data frame summarizing the importance of each environmental variable, including `%IncMSE` (percent increase in mean squared error) and `IncNodePurity` (increase in node purity).
  \item `p`: A ggplot2 bar plot visualizing the importance of each environmental variable based on `%IncMSE`.
}
}
\description{
The `ramdom_env.plot` function uses Random Forest analysis to identify the importance of environmental variables in explaining the variation along the first NMDS axis (MDS1) derived from Bray-Curtis distances of OTU data.
}
\details{
The function performs the following steps:
\enumerate{
  \item Converts OTU data to a data frame and calculates Bray-Curtis distances.
  \item Computes Non-Metric Multidimensional Scaling (NMDS) using Bray-Curtis distances.
  \item Combines the first NMDS axis (MDS1) with environmental variables to create a model dataset.
  \item Applies Random Forest regression to assess the importance of each environmental variable in explaining MDS1 variation.
  \item Outputs a bar plot of variable importance based on `%IncMSE`.
}
}
\examples{
\dontrun{

result <- ramdom_env.plot(ps = ps, env = env, seed = 42)
}

Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
