% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomforest.metm.R
\name{randomforest.metm}
\alias{randomforest.metm}
\title{Random forest modeling for microbiome data}
\usage{
randomforest.metm(
  otu = NULL,
  tax = NULL,
  map = NULL,
  tree = NULL,
  ps = NULL,
  group = "Group",
  optimal = 20,
  rfcv = FALSE,
  nrfcvnum = 5,
  min = -1,
  max = 5,
  fill = "Phylum",
  lab = "id"
)
}
\arguments{
\item{otu}{OTU/ASV table;}

\item{tax}{taxonomy table;}

\item{map}{Sample metadata;}

\item{ps}{phyloseq object of microbiome;}

\item{optimal}{important OTU number which selected}

\item{rfcv}{TURE or FELSE,whether need to do cross-validation}

\item{nrfcvnum}{Number of cross-validation}

\item{min}{Circle diagram inner diameter adjustment}

\item{max}{Circle diagram outer diameter adjustment}

\item{Group}{column name for groupID in map table.}
}
\value{
list contain ggplot object and table.
}
\description{
Random forest modeling for microbiome data
}
\examples{
result = randomforest.metm(ps = ps,group  = "Group",optimal = 20,rfcv = TRUE,nrfcvnum = 5,min = -1,max = 5)
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
