% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomforest.omics.R
\name{randomforest.omics}
\alias{randomforest.omics}
\title{Random forest modeling for microbiome data}
\usage{
randomforest.omics(
  ps = NULL,
  group = "Group",
  optimal = 20,
  rfcv = FALSE,
  nrfcvnum = 5,
  min = -1,
  max = 5,
  fill = "filed",
  lab = "id"
)
}
\arguments{
\item{ps}{phyloseq object of microbiome}

\item{optimal}{important OTU number which selected}

\item{rfcv}{TURE or FELSE,whether need to do cross-validation}

\item{nrfcvnum}{Number of cross-validation}

\item{min}{Circle diagram inner diameter adjustment}

\item{max}{Circle diagram outer diameter adjustment}

\item{Group}{column name for groupID in map table.}
}
\value{
list contain ggplot object and table.
}
\description{
Random forest modeling for microbiome data
}
\examples{
result = randomforest.omics(ps = ps03,group  = "Group",optimal = 20,rfcv = TRUE,nrfcvnum = 5,min = -1,max = 5)
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
