% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomforest.trans.R
\name{randomforest.trans}
\alias{randomforest.trans}
\title{Random forest modeling for transcriptome functional composition data}
\usage{
randomforest.trans(
  otu = NULL,
  tax = NULL,
  map = NULL,
  tree = NULL,
  ps = NULL,
  group = "Group",
  optimal = 20,
  rfcv = FALSE,
  nrfcvnum = 5,
  min = -1,
  max = 5,
  fill = "Level1",
  lab = "id"
)
}
\arguments{
\item{otu}{Transcriptome functional composition table.}

\item{tax}{transcriptome functional classification table.}

\item{map}{Sample metadata.}

\item{tree}{Phylogenetic tree.}

\item{ps}{A phyloseq format file used as an alternative for the input containing transcriptome functional composition data, taxonomy table, and sample metadata.}

\item{group}{Column name for groupID in map table(sample metadata).}

\item{optimal}{Select the number of important genes to be displayed.}

\item{rfcv}{TURE or FALSE,whether need to do cross-validation.}

\item{nrfcvnum}{Number of cross-validation.}

\item{min}{Circle diagram inner diameter adjustment.}

\item{max}{Circle diagram outer diameter adjustment.}
}
\value{
list contain ggplot object and table:
\item{p1}{A lollipop diagram of the selected number of important genes.}
\item{p2}{A circle diagram of the selected number of important genes.}
\item{a3}{Taxonomic annotation information for the selected number of important genes and their importance.}
\item{pn}{Combined table of model accuracy rates and confusion matrix.}
\item{a2}{Taxonomic annotation information for all genes and their importance.}
}
\description{
Random forest modeling for transcriptome functional composition data was used to screen for characteristic genes.
}
\examples{
library(dplyr)
library(ggClusterNet)
library(caret)
library(randomForest)
# Example with a phyloseq object
data(ps.trans)
ps = ps.trans\%>\% filter_taxa(function(x) sum(x ) > 5 , TRUE)
res <- randomforest.trans( ps= ps, group  = "Group", optimal = 50)
p1 = res[[1]];p1
p2=res[[2]];p2
p3=res[[4]];p3
dat1 =res[[3]];dat1
dat2 =res[[5]];dat2
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njau.edu.cn}
}
