% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefy_matrix.R
\name{rarefy_matrix}
\alias{rarefy_matrix}
\title{Rarefy a count matrix}
\usage{
rarefy_matrix(mat, N)
}
\arguments{
\item{mat}{numeric matrix of counts}

\item{N}{integer, rarefaction depth}
}
\value{
numeric matrix of rarefied counts
}
\description{
Rarefy the OTU/feature matrix to depth N.
}
