% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA_CCA.R
\name{RDA_CCA}
\alias{RDA_CCA}
\title{Perform Redundancy Analysis (RDA) or Canonical Correspondence Analysis (CCA)}
\usage{
RDA_CCA(ps = NULL, env = env, group = "Group", chose.env = FALSE)
}
\arguments{
\item{ps}{A `phyloseq` object containing microbial community data, including OTU table and metadata.}

\item{env}{A data frame of environmental variables, where rows correspond to samples.}

\item{group}{A character string specifying the column in `ps` metadata used for grouping. Default is `"Group"`.}

\item{chose.env}{A logical value indicating whether to perform stepwise selection of environmental variables. Default is `FALSE`.}
}
\description{
The `RDA_CCA` function performs Redundancy Analysis (RDA) or Canonical Correspondence Analysis (CCA)
on microbial community data with associated environmental factors. It automatically selects the
appropriate method based on Detrended Correspondence Analysis (DCA) results and allows for stepwise
environmental variable selection to reduce collinearity.
}
\details{
The function proceeds as follows:
\enumerate{
  \item Normalizes the OTU table to relative abundances.
  \item Aligns environmental data with microbial data based on sample IDs.
  \item Uses DCA to decide whether to apply RDA (if DCA1 < 3) or CCA (if DCA1 > 4).
  \item Optionally performs stepwise variable selection to remove collinear environmental variables based on VIF.
  \item Performs RDA or CCA and computes significance values using permutation tests.
  \item Generates a biplot with sample and environmental variable coordinates.
}
}
\examples{
\dontrun{
result <- RDA_CCA(ps = ps, env = env, group = "Group" )
}
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
