% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA_CCA_explain_percent.R
\name{RDA_CCA_explain_percent}
\alias{RDA_CCA_explain_percent}
\title{Calculate and Visualize Environmental Contributions in RDA/CCA}
\usage{
RDA_CCA_explain_percent(ps = ps, env.dat = envRDA)
}
\arguments{
\item{ps}{A `phyloseq` object containing microbial community data, including the OTU table and sample metadata.}

\item{env.dat}{A data frame of environmental variables, where rows correspond to samples and columns represent variables.}
}
\value{
A list containing:
\itemize{
  \item `out`: A data frame summarizing the percentage contribution of each environmental variable to the constrained variance.
  \item `exp`: A data frame summarizing the total explained and unexplained variance as proportions.
  \item `p`: A ggplot2 object visualizing the percentage contributions of environmental variables as a horizontal bar plot.
}
}
\description{
The `RDA_CCA_explain_percent` function performs Redundancy Analysis (RDA) or Canonical Correspondence Analysis (CCA)
to evaluate the contributions of environmental variables to microbial community variation. The function calculates
the explained and unexplained variance and generates a bar plot to visualize the contribution of individual variables.
}
\examples{
\dontrun{
library(phyloseq)
library(vegan)
library(ggplot2)
library(tidyr)
result <- RDA_CCA_explain_percent(ps = ps, env.dat = envRDA)

# Extract results
out <- result[[1]]  # Environmental variable contributions
exp <- result[[2]]  # Explained and unexplained variance
plot <- result[[3]] # Bar plot
}
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
