% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdacca.hp.micro.R
\name{rdacca.hp.micro}
\alias{rdacca.hp.micro}
\title{Hierarchical Partitioning for RDA, CCA, and db-RDA}
\usage{
rdacca.hp.micro(OTU = mite, env = env1, cca = FALSE)
}
\arguments{
\item{OTU}{A data frame or matrix of OTU abundance data, where rows represent samples and columns represent OTUs.}

\item{env}{A data frame of environmental variables, where rows correspond to samples and columns are variables.}

\item{cca}{A logical value indicating whether to perform hierarchical partitioning in canonical correspondence analysis (CCA). Default is `FALSE`.}
}
\value{
A list containing:
\itemize{
  \item `p`: A ggplot2 object visualizing the hierarchical partitioning results for RDA.
  \item `dat`: A data frame containing hierarchical partitioning results for RDA or CCA.
  \item `p1`: A ggplot2 object visualizing the hierarchical partitioning results for db-RDA.
  \item `dat2`: A data frame containing hierarchical partitioning results for db-RDA.
}
}
\description{
The `rdacca.hp.micro` function performs hierarchical partitioning to decompose the variance explained by environmental variables in redundancy analysis (RDA), canonical correspondence analysis (CCA), and distance-based redundancy analysis (db-RDA). It provides visualizations of individual contributions and outputs detailed hierarchical partitioning results.
}
\details{
The function performs the following steps:
\enumerate{
  \item **RDA (Redundancy Analysis)**:
    - Standardizes the OTU data using Hellinger transformation.
    - Conducts RDA and calculates the adjusted R² for each environmental variable using hierarchical partitioning.
    - Visualizes the individual contributions of environmental variables in a bar plot.
  \item **CCA (Canonical Correspondence Analysis)** (if `cca = TRUE`):
    - Uses the original OTU abundance matrix without transformation.
    - Performs CCA and calculates the adjusted R² for each environmental variable.
    - Visualizes the contributions of environmental variables in a bar plot.
  \item **db-RDA (Distance-based RDA)**:
    - Computes Bray-Curtis distances from the OTU data.
    - Performs db-RDA and calculates the adjusted R² for each environmental variable using hierarchical partitioning.
    - Visualizes the contributions of environmental variables in a bar plot.
}
}
\examples{
\dontrun{
res = rdacca.hp.micro(OTU = ps.tem \%>\% filter_OTU_ps(500) \%>\%vegan_otu() \%>\% as.data.frame(), env = ftab[,1:5], cca = FALSE)
p3_rda = res[[1]]
dat1 = res[[2]]
p3_db_rda  = res[[3]]
dat2 = res[[4]]
}
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
