% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.show.ms.R
\name{reaction.show.ms}
\alias{reaction.show.ms}
\title{Perform reaction pathway enrichment analysis}
\usage{
reaction.show.ms(ps, dif.method = "wilcox")
}
\arguments{
\item{ps}{A phyloseq format file used as an alternative for the input containing metabolite composition table, metabolite classification table, and sample metadata.}

\item{dif.method}{The differential method to use for analysis (default is "wilcox").}
}
\value{
A list containing the following components:
\describe{
  \item{Plots}{List of ggplot objects representing the reaction pathway enrichment analysis plots.}
  \item{Plotdata}{List of data frames containing reaction pathway enrichment analysis results.}
  }
}
\description{
This function conducts reaction pathway enrichment analysis based on the specified differential method (wilcox, t.test, etc.) for a phyloseq format file.
}
\examples{
\dontrun{
ps.ms3 = ps.ms \%>\% tax_glom_wt("KEGGID")
res3 = reaction.show.ms(ps = ps.ms3, dif.method = "wilcox")
res3$Plots$OE.WT.plot
dat = res3$Plotdata$OE.WT
head(dat)
}
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njau.edu.cn}
}
