% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_tree.micro.R
\name{remove_rankID}
\alias{remove_rankID}
\title{Generate Phylogenetic Tree with Taxonomic and Interaction Annotations}
\usage{
remove_rankID(taxtab)
}
\arguments{
\item{ps}{A `phyloseq` object containing microbiome data.}

\item{Top}{An integer specifying the number of top OTUs to include based on abundance. Default is `100`.}
}
\value{
A list containing:
\describe{
  \item{p0}{The inward circular phylogenetic tree with basic taxonomic information.}
  \item{p1}{The inward circular tree with taxonomic annotations and OTU points.}
  \item{p2}{The tree with OTU interaction links visualized as edges.}
  \item{p3}{The tree with OTU abundance visualized as stars.}
  \item{p4}{The tree with OTU abundance and OTU labels.}
  \item{p5}{The tree with sample group-level bar plots.}
  \item{p6}{The circular tree with taxonomic annotations and OTU points.}
  \item{p7}{The circular tree with sample group-level bar plots.}
  \item{data}{A data frame containing tip point annotations, including taxonomic levels and OTU names.}
}
}
\description{
This function generates a detailed phylogenetic tree annotated with taxonomic information, OTU interactions, and sample group-level bar plots. The tree can be visualized in both circular and inward circular layouts, with various layers of data added, including correlations, bar plots, and abundance data.
}
\details{
The function performs the following steps:
\itemize{
  \item Extracts and processes taxonomic data from the `phyloseq` object.
  \item Constructs a phylogenetic tree using the `MicrobiotaProcess::convert_to_treedata` function.
  \item Annotates the tree with taxonomic information at the specified levels.
  \item Adds OTU interaction links to the tree based on correlation analysis using the `ggClusterNet::corMicro` function.
  \item Visualizes OTU abundance with star glyphs.
  \item Adds sample group-level bar plots to the tree for visualizing the mean abundance of OTUs across groups.
}
}
\examples{
\dontrun{
result <- phy_tree.micro(ps = ps.16s,Top = 100)
p0 = result[[1]]
}

}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
