% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfcv.Metf.R
\name{rfcv.metf}
\alias{rfcv.metf}
\title{For cross-validation of metagenome functional composition data}
\usage{
rfcv.metf(
  otu = NULL,
  tax = NULL,
  map = NULL,
  tree = NULL,
  ps = NULL,
  group = "Group",
  optimal = 20,
  nrfcvnum = 5
)
}
\arguments{
\item{otu}{Metagenome functional composition table.}

\item{tax}{taxonomy table.}

\item{map}{Sample metadata.}

\item{ps}{A phyloseq format file used as an alternative for the input containing metagenome functional composition table, tax, and sample metadata.}

\item{optimal}{Important gene number which selected to plot.}

\item{nrfcvnum}{Number of cross-validation.}

\item{Group}{Column name for groupID in map table.}

\item{rfcv}{TURE or FELSE,whether need to do cross-validation.}
}
\value{
list contain ggplot object and table.
}
\description{
For cross-validation of metagenome functional composition data
}
\examples{
ps =ps.kegg \%>\% filter_OTU_ps(Top = 1000)
result =rfcv.metf(ps = ps \%>\% filter_OTU_ps(200),group  = "Group",optimal = 20,nrfcvnum = 6)
prfcv = result[[1]]
prfcv
rfcvtable = result[[3]]
rfcvtable
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn},  Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
