% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfcv.micro.R
\name{rfcv.Micro}
\alias{rfcv.Micro}
\title{For cross-validation of microbiome data}
\usage{
rfcv.Micro(
  otu = NULL,
  tax = NULL,
  map = NULL,
  tree = NULL,
  ps = NULL,
  group = "Group",
  optimal = 20,
  nrfcvnum = 5
)
}
\arguments{
\item{otu}{OTU/ASV table;}

\item{tax}{taxonomy table}

\item{map}{Sample metadata;}

\item{ps}{phyloseq object of microbiome}

\item{optimal}{important OTU number which selected}

\item{nrfcvnum}{Number of cross-validation}

\item{Group}{column name for groupID in map table.}

\item{rfcv}{TURE or FELSE,whether need to do cross-validation}
}
\value{
list contain ggplot object and table.
}
\description{
For cross-validation of microbiome data
}
\examples{
# data form github
result = Micro.rfcv(otu = NULL,tax = NULL,map = NULL,tree = NULL ,ps = ps_rela,group  = "Group",optimal = 20,nrfcvnum = 6)
prfcv = result[[1]]# plot rfcv
rfcvtable = result[[3]]# table rfcv
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn},  Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
