% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfcv.ms.R
\name{rfcv.ms}
\alias{rfcv.ms}
\title{For cross-validation of metabolite data}
\usage{
rfcv.ms(
  otu = NULL,
  tax = NULL,
  map = NULL,
  tree = NULL,
  ps = NULL,
  group = "Group",
  optimal = 20,
  nrfcvnum = 5
)
}
\arguments{
\item{otu}{metabolite composition table.}

\item{tax}{taxonomy table.}

\item{map}{Sample metadata.}

\item{ps}{phyloseq object of metabolite.}

\item{optimal}{important metabolite number which selected.}

\item{nrfcvnum}{Number of cross-validation.}

\item{Group}{column name for groupID in map table.}

\item{rfcv}{TURE or FELSE,whether need to do cross-validation.}
}
\value{
list contain ggplot object and table.
}
\description{
For cross-validation of metabolite data.
}
\examples{
library(dplyr)
ps_rela <- ps.ms\%>\%phyloseq::transform_sample_counts(function(x) {x/sum(x)} )
result = rfcv.ms(otu = NULL,tax = NULL,map = NULL,tree = NULL ,ps = ps_rela,group  = "Group",optimal = 20,nrfcvnum = 6)
prfcv = result[[1]]# plot rfcv
rfcvtable = result[[3]]# table rfcv
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn},  Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
