% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfcv.omics.R
\name{rfcv.omics}
\alias{rfcv.omics}
\title{For cross-validation of multi-omics data}
\usage{
rfcv.omics(ps = NULL, group = "Group", optimal = 20, nrfcvnum = 5)
}
\arguments{
\item{ps}{phyloseq object of multi-omics data}

\item{optimal}{important OTU number which selected}

\item{nrfcvnum}{Number of cross-validation}

\item{Group}{column name for groupID in map table.}

\item{rfcv}{TURE or FELSE,whether need to do cross-validation}
}
\value{
list contain ggplot object and table.
}
\description{
For cross-validation of multi-omics data
}
\examples{
# data form github
result = rfcv.omics(ps = ps,group  = "Group",optimal = 20,nrfcvnum = 6)
prfcv = result[[1]]# plot rfcv
rfcvtable = result[[3]]# table rfcv
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn},  Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
