% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roc.metf.R
\name{Roc.metf}
\alias{Roc.metf}
\title{Comparison of three machine methods (randomforest,SVM,GLM)}
\usage{
Roc.metf(
  otu = NULL,
  tax = NULL,
  map = NULL,
  tree = NULL,
  ps = NULL,
  group = "Group",
  repnum = 5
)
}
\arguments{
\item{otu}{Metagenome functional composition table.}

\item{tax}{taxonomy table.}

\item{map}{Sample metadata.}

\item{ps}{A phyloseq format file used as an alternative for the input containing metagenome functional composition table, tax, and sample metadata.}

\item{repnum}{Modeling times.}

\item{Group}{Column name for groupID in map table.}
}
\value{
list contain ggplot object and table.
}
\description{
Comparison of three machine methods (randomforest,SVM,GLM).
}
\examples{
id = sample_data(ps)$Group \%>\% unique()
aaa = combn(id,2)
i= 1
group = c(aaa[1,i],aaa[2,i])
pst = ps \%>\% subset_samples.wt("Group",group) \%>\%filter_taxa(function(x) sum(x ) > 10, TRUE)
res = Roc.metf( ps = pst,group  = "Group",repnum = 5)
p33.1 =  res[[1]]
p33.1
p33.2 =  res[[2]]
p33.2
dat =  res[[3]]
dat
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn},  Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
