% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roc.omics.R
\name{Roc.omics}
\alias{Roc.omics}
\title{Comparison of ROC curves of three machine methods (randomforest,SVM,GLM).}
\usage{
Roc.omics(ps = NULL, group = "Group", repnum = 5)
}
\arguments{
\item{ps}{phyloseq object of microbiome.}

\item{repnum}{The number of folds for cross-validation.}

\item{Group}{column name for groupID in map table.}
}
\value{
A list object including the following components:
\item{g2}{Final ROC curve plot of three machine methods.}
\item{AUC}{AUC values for each classification model.}
\item{df}{Combined dataframe with true and predicted values}
\item{rocdata list}{List of ROC curve data for Random Forest, SVM, and GLM}
}
\description{
Comparison of three machine methods (randomforest,SVM,GLM), and
 these model was evaluated using k-fold cross-validation.
}
\examples{
library(caret)
library(randomForest)
library(ROCR)
library(e1071)
id = sample_data(ps03)$Group \%>\% unique()
aaa = combn(id,2)
i= 1
group = c(aaa[1,i],aaa[2,i])
pst =ps03 \%>\% subset_samples.wt("Group",group)
res = Roc.omics( ps = pst \%>\% filter_OTU_ps(1000),group  = "Group",repnum = 5)
p33.1 =  res[[1]]
p33.1
p33.2 =  res[[2]]
p33.2
dat =  res[[3]]
dat
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
