% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_metabolite_selection.R
\name{sample_size_constraint}
\alias{sample_size_constraint}
\title{Sample Size Constraint}
\usage{
sample_size_constraint(scores, sample_size, results = NULL)
}
\arguments{
\item{scores}{Numeric vector of metabolite scores}

\item{sample_size}{Integer, number of samples}

\item{results}{Results object (for inferring sample size if not provided)}
}
\value{
Integer, maximum recommended number of metabolites
}
\description{
Determines maximum number of features based on sample size to avoid overfitting.
}
\keyword{internal}
