% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.m.Group.micro.R
\name{sankey.m.Group.micro}
\alias{sankey.m.Group.micro}
\title{Generate Group-Level Taxonomic Sankey Diagram for Microbial Data}
\usage{
sankey.m.Group.micro(ps = ps, rank = 6, Top = 50)
}
\arguments{
\item{ps}{A `phyloseq` object containing microbiome data.}

\item{rank}{A numeric value specifying the taxonomic rank for aggregation. Default is `6` (e.g., Genus level).}

\item{Top}{An integer specifying the number of top OTUs to include based on abundance. Default is `50`.}
}
\value{
A list containing:
\describe{
  \item{An interactive Sankey diagram generated with `networkD3::sankeyNetwork`.}
  \item{A data frame containing the Sankey diagram links and node data.}
}
}
\description{
This function creates group-level Sankey diagrams to visualize taxonomic transitions from higher to lower taxonomic levels (e.g., Kingdom to Genus).
It generates interactive Sankey diagrams for grouped microbial data based on the specified taxonomic rank and top OTUs.
}
\details{
The function performs the following steps:
\itemize{
  \item Extracts OTU and taxonomic data from the `phyloseq` object.
  \item Aggregates taxonomic information up to the specified rank and filters the top `Top` OTUs by abundance.
  \item Calculates mean abundance values for each taxonomic level within groups.
  \item Builds source-target relationships for taxonomic transitions between levels (e.g., Kingdom → Phylum, Phylum → Class).
  \item Generates an interactive Sankey diagram visualizing taxonomic transitions and their relative abundances at the group level.
}

The Sankey diagram is interactive and allows users to explore taxonomic transitions and their abundances across groups.
}
\examples{
\dontrun{
result = sankey.m.Group.micro(ps = ps.16s  \%>\% subset_taxa.wt("Species","Unassigned",TRUE),rank = 6,Top = 50)
p = result[[1]]
dat = result[[2]]
 }

}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
