% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_IS.ms.R
\name{scale_IS.ms}
\alias{scale_IS.ms}
\title{Scale metabolome data by normalizing based on selected metabolite(s)}
\usage{
scale_IS.ms(ps = ps.ms, IS = "metab_8497")
}
\arguments{
\item{ps}{A phyloseq format file used as an alternative for the input containing metabolite, tax, and sample data.}

\item{IS}{A character vector specifying the selected metabolite(s).}
}
\value{
The scaled phyloseq format file after normalization based on the selected metabolite(s).
}
\description{
The scale_IS.ms function scales  metabolite data by normalizing based on selected metabolite(s).
It  standardizes the  metabolite data of each metabolite  by the expression levels of selected metabolite(s) in each sample.
}
\examples{
ps.ms2 = scale_IS.ms(ps = ps.ms,IS = "metab_8497")
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
