% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_QC.ms.R
\name{scale_QC.ms}
\alias{scale_QC.ms}
\title{Scale  metabolite data by normalizing based on Quality Control (QC) samples}
\usage{
scale_QC.ms(ps = ps.ms, QC = c("OE36_2", "OE36_3", "OE2_1"))
}
\arguments{
\item{ps}{A phyloseq format file used as an alternative for the input containing metabolite, tax, and sample metadata.}

\item{QC}{A character vector specifying the QC sample names.}
}
\value{
The scaled phyloseq format file after normalization based on the QC samples.
}
\description{
The scale_QC.ms function scales metabolite data by normalizing based on Quality Control (QC) samples.
It calculates the average value of each metabolite in the QC samples and standardizes  metabolite data of each sample by this average value.
}
\examples{
ps.ms2 = scale_QC.ms(ps = ps.ms,QC = c("OE36_2","OE36_3","OE2_1"))
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
