% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schur.R
\name{schur}
\alias{schur}
\title{Schur Product (Element-wise multiplication)

Calculates the element-wise product (Hadamard product) of two matrices.}
\usage{
schur(a, b)
}
\arguments{
\item{a}{numeric matrix}

\item{b}{numeric matrix}
}
\value{
numeric matrix resulting from element-wise multiplication
}
\description{
Schur Product (Element-wise multiplication)

Calculates the element-wise product (Hadamard product) of two matrices.
}
\keyword{internal}
