% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statSuper.R
\name{statSuper}
\alias{statSuper}
\title{Perform differential statistical analysis on metabolites data}
\usage{
statSuper(
  otu = NULL,
  tax = NULL,
  map = NULL,
  tree = NULL,
  ps = NULL,
  group = "Group",
  pvalue = 0.05,
  artGroup = NULL,
  method = "wilcox"
)
}
\arguments{
\item{otu}{A data frame containing metabolite data. If `NULL`, the `ps` object is used.}

\item{tax}{A data frame containing taxonomy data. If `NULL`, the `ps` object is used.}

\item{map}{A data frame containing sample metadata. If `NULL`, the `ps` object is used.}

\item{tree}{A phylogenetic tree object. If `NULL`, the `ps` object is used.}

\item{ps}{A phyloseq format file used as an alternative for the input containing metabolite composition table,
metabolite classification table, and sample metadata.}

\item{group}{Column name in the mapping data specifying group information. Default is "Group".}

\item{pvalue}{Threshold p-value for differential analysis. Default is 0.05.}

\item{artGroup}{User-defined group combinations.}

\item{method}{Statistical test method for differential analysis. Default is "wilcox" (Wilcoxon test).}
}
\value{
Data frame containing statistical results, differential abundance, and relative abundance information.
}
\description{
This function conducts statistical analysis on metabolites data, including
differential abundance analysis and relative abundance calculations.
}
\examples{
result1 = statSuper(ps = ps.ms,group  = "Group",artGroup = NULL,method = "wilcox")
head(result1)
result2 = statSuper(ps = ps.ms,group  = "Group",artGroup = NULL,method = "ttext")
head(result2)
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
