% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_diff.micro.R
\name{stemp_diff.micro}
\alias{stemp_diff.micro}
\title{Statistical Comparison of Taxonomic Abundances Between Groups}
\usage{
stemp_diff.micro(
  ps = ps,
  Top = 20,
  ranks = 6,
  method = "TMM",
  test.method = "t.test"
)
}
\arguments{
\item{ps}{A `phyloseq` object containing microbiome data (OTU table, taxonomic table, and sample metadata).}

\item{Top}{An integer specifying the maximum number of OTUs or taxa to analyze. Default is `20`.}

\item{ranks}{A character or numeric value specifying the taxonomic rank to analyze. For example: `"Genus"`, `"Family"`, or a numeric rank index. Default is `6`.}

\item{method}{A character string indicating the normalization method. Options are `TMM` (default), `RLE`, or others supported by the `edgeR` package.}

\item{test.method}{A character string specifying the statistical test. Options are `"t.test"` (default) for Student's t-test or `"wilcox.test"` for the Wilcoxon rank-sum test.}
}
\value{
A composite plot (`patchwork` object) consisting of three panels:
\itemize{
  \item A bar plot showing mean proportions of taxa between groups.
  \item A confidence interval plot showing the difference in mean proportions with 95% confidence intervals.
  \item A column of adjusted p-values associated with the taxa.
}
}
\description{
The `stemp_diff.micro` function identifies significant differences in taxonomic abundances between two groups
using statistical tests such as `t-test` or `Wilcoxon test`. It visualizes the results as bar plots and confidence interval plots.
}
\details{
The function performs the following steps:
\itemize{
  \item Aggregates the data at the specified taxonomic rank.
  \item Normalizes the data using the specified method (`TMM` by default).
  \item Filters taxa to include only the top `Top` most abundant taxa.
  \item Applies the specified statistical test (`t.test` or `wilcox.test`) for each taxon to compare mean proportions between groups.
  \item Adjusts the p-values using the Bonferroni correction method.
  \item Creates a composite plot visualizing the mean proportions, confidence intervals, and adjusted p-values.
}
}
\examples{
\dontrun{
map = phyloseq::sample_data(ps.16s)
allgroup <- combn(unique(map$Group),2)
plot_list <- list()
for (i in 1:dim(allgroup)[2]) {
 ps_sub <- phyloseq::subset_samples(ps.16s,Group \%in\% allgroup[,i]);ps_sub
 p <- stemp_diff.micro(ps = ps_sub,Top = 20,ranks = 6)
 plot_list[[i]] <- p}
}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
