% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_ps_microbe_type.R
\name{subset_main_kingdom_ps}
\alias{subset_main_kingdom_ps}
\title{Subset phyloseq object by dominant kingdom}
\usage{
subset_main_kingdom_ps(ps, kingdom_col = "Kingdom", min_major_frac = 0.5)
}
\arguments{
\item{ps}{A phyloseq object.}

\item{kingdom_col}{tax_table 中 Kingdom 对应的列名，默认 "Kingdom"。}

\item{min_major_frac}{传给 guess_ps_microbe_type，用于类型判断（这里只用于理由说明）。}
}
\value{
一个新的 phyloseq 对象（按主 Kingdom 过滤后的子集）。
}
\description{
使用 guess_ps_microbe_type 的结果，自动选择数量最多的 Kingdom，
然后用 subset_taxa.wt() 按该 Kingdom 过滤 OTU/ASV。
- 如果是细菌数据，会选出 Bacteria 对应的所有条目；
- 真菌数据类似；
- 如果是混合数据，则选计数最多的那个 Kingdom；
- 如果没有 Kingdom 信息，则直接返回原始 ps。
}
