% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_feature_selection.R
\name{summary.ms_feature_selection}
\alias{summary.ms_feature_selection}
\title{Summary Method for ms_feature_selection Objects}
\usage{
\method{summary}{ms_feature_selection}(object, ...)
}
\arguments{
\item{object}{Object of class ms_feature_selection}

\item{...}{Additional arguments (not used)}
}
\description{
Summary Method for ms_feature_selection Objects
}
