% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.metf.R
\name{t.metf}
\alias{t.metf}
\title{Differential abundance analysis using t test for metagenome functional data}
\usage{
\method{t}{metf}(
  ps = ps,
  artGroup = NULL,
  group = "Group",
  pvalue = 0.05,
  FCvalue = 2,
  padjust = TRUE,
  padjust_method = "BH",
  log2FC_threshold = 10,
  ncol = 3,
  nrow = 1,
  outpath = NULL
)
}
\arguments{
\item{ps}{A phyloseq format file used as an alternative for the input containing metagenome
functional composition table, tax, and sample metadata.}

\item{artGroup}{A matrix specifying custom group comparisons. Optional,
the standard format is two rows,Each column represents the two groups selected for comparison,
with the first row defaulting to the reference control group.}

\item{group}{A string specifying the grouping variable in the sample metadata. Default is `"Group"`.}

\item{pvalue}{A numeric value specifying the significance threshold for p value. Default is `0.05`.}

\item{FCvalue}{A numeric value specifying the significance threshold for fold change value. Default is `2`.}

\item{padjust}{A logical value that is used to determine whether to perform p-value correction. Default is `TRUE`.}

\item{padjust_method}{Select the method for p-value correction("holm", "hochberg", "hommel", "bonferroni", "BH", "BY",
"fdr"),when padjust is `TRUE`. Default is `BH`.}

\item{log2FC_threshold}{A numeric value specifying the plot threshold for
log2fold_change_value. Default is `10`.}

\item{ncol}{Number of columns for arranging plots.Default is `3`.}

\item{nrow}{Number of rows for arranging plots.Default is `1`.}

\item{outpath}{Path to save the analysis results and plots.}
}
\value{
A list containing the following components:
\item{Volcano_all_plot}{A combined volcano plots of all comparison groups.}
\item{result_all}{Data frame containing all comparison groups analysis result.}
\item{meta}{Data frame containing the union and intersection of differentially expressed genes across all pairwise comparison groups.}
\item{plot_list}{A list of individual volcano plots.}
}
\description{
This function performs differential abundance analysis for metagenome functional data using the `t.test` function.
It identifies differentially abundant taxa betwee two groups in a `phyloseq` object or selected comparison groups,
visualizes the results with volcano plots, and outputs realtive abundance data with test results.
}
\examples{
\dontrun{
ps=ps.kegg \%>\% filter_OTU_ps(Top = 1000)
res = t.metf(ps = ps,group  = "Group",artGroup =NULL,group  = "Group",
             pvalue=0.05,FCvalue=2,padjust=TRUE,padjust_method="BH",
             log2FC_threshold=10,ncol=3,nrow=1,outpath=NULL)
p17=res[[1]]
p17
data=res[[2]]
head(data)
inter_union=res[[3]]
head(inter_union)
p17.1 = res [[4]][1]
p17.1
p17.2 = res [[4]][2]
p17.2
p17.3 = res [[4]][3]
p17.3
}
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njau.edu.cn}
}
