% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_my.R
\name{theme_my}
\alias{theme_my}
\title{Custom Theme and Color Settings for the analyse}
\usage{
theme_my(ps = ps, gnum = NULL)
}
\arguments{
\item{ps}{A `phyloseq` object containing the microbial community data and metadata.}

\item{gnum}{Calculate the Number of Unique Groups in Sample Metadata}
}
\description{
The `theme_my` function provides two custom ggplot2 themes (`mytheme1` and `mytheme2`) and dynamically assigns color palettes based on the number of unique groups (`gnum`) in a phyloseq object. This function is specifically designed for creating consistent and aesthetically pleasing visualizations of microbial community data.
}
\examples{
res = theme_my(ps.16s)
mytheme1 = res[[1]]
mytheme2 = res[[2]]
colset1 = res[[3]]

}
\author{
Tao Wen \email{2018203048@njau.edu.cn}, Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
