% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_mkegg.R
\name{trans_mkegg}
\alias{trans_mkegg}
\title{KEGG Pathway Transformation using Mkegg}
\usage{
trans_mkegg(ps.trans = ps.trans)
}
\arguments{
\item{ps.trans}{A Phyloseq object containing OTU table and taxonomic data.
The default is `ps.trans` which assumes that the Phyloseq object is provided externally.}
}
\value{
A transformed Phyloseq object where the OTU table is aggregated by KEGG pathway IDs
and taxonomic information is updated with KEGG pathway descriptions.
}
\description{
This function downloads and processes KEGG pathway data using the `Mkegg` option
for KEGG analysis from `clusterProfiler`. It maps KEGG Orthology (KO) terms to
KEGG pathways and integrates the data with OTU tables from a Phyloseq object.
The result is a transformed Phyloseq object with pathway-level information.
}
\examples{
\dontrun{
ps_transformed <- trans_mkegg(ps.trans)
}
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
