% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ven.network.metm.R
\name{ven.network.metm}
\alias{ven.network.metm}
\title{Generate a Circular Microbial Network}
\usage{
ven.network.metm(ps = ps, N = 0.5, fill = "Phylum")
}
\arguments{
\item{ps}{A `phyloseq` object containing microbiome data.}

\item{N}{Numeric. A threshold for selecting nodes in the network. Default is `0.5`.}

\item{fill}{A character string specifying the taxonomic rank used to color the nodes. Default is `"Phylum"`.}
}
\value{
A list containing:
\describe{
  \item{plot}{A `ggplot2` object representing the microbial network.}
  \item{plotdata}{A data frame containing node coordinates and associated metadata.}
}
}
\description{
This function generates a circular microbial network based on microbial abundance data in a `phyloseq` object.
It uses the `ggClusterNet` package to compute relationships between operational taxonomic units (OTUs)
and visualizes the network with sample-level metadata.
}
\details{
This function builds a microbial network using a circular layout. OTUs are placed hierarchically based on their
abundance and relationships with samples. The edges represent connections between OTUs and sample groups,
while nodes represent microbial taxa or samples.
}
\examples{
result = ven.network.metm(ps = ps.16s,N = 0.5,fill = "Phylum")
p14  = result[[1]]
p14
dat = result[[2]]
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Peng-Hao Xie \email{2019103106@njqu.edu.cn}
The function includes several steps:
\itemize{
  \item Calculates the microbial relationships using `ggClusterNet::div_network`.
  \item Uses a progressive circle packing algorithm to position OTUs and groups.
  \item Combines abundance data and taxonomic annotations for visualization.
  \item Generates a `ggplot2` plot showing the network with colored nodes and edges.
}
}
