% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VenSeper.metf.R
\name{VenSuper.metf}
\alias{VenSuper.metf}
\title{Generate Venn Diagrams and Analyze KEGG Groups}
\usage{
VenSuper.metf(ps = NULL, group = "Group", num = 6, j = j)
}
\arguments{
\item{ps}{A `phyloseq` object containing kegg data.}

\item{group}{A character string specifying the grouping variable in the sample metadata. Default is `"Group"`.}

\item{num}{An integer specifying the threshold for selecting kegg in Venn groups. Default is `6`.}
}
\value{
A list containing the following elements:
\describe{
  \item{pa}{A combined bar plot visualization of kegg for each Venn segment.}
  \item{pb}{A combined box plot visualization for the selected kegg in each segment.}
  \item{pc}{A combined relative abundance bar plot for each Venn segment.}
  \item{dat.f}{A list of data frames containing bar plot data for each Venn segment.}
  \item{dat.f2}{A list of data frames containing statistical analysis results for each Venn segment.}
}
}
\description{
This function creates Venn diagrams to compare microbial groups based on presence/absence thresholds,
and performs detailed analyses for each group. The results include Venn diagrams, bar plots, and
boxplots for visualizing group-specific KEGG data.
}
\details{
The function performs the following steps:
\itemize{
  \item Extracts OTU presence/absence data for each microbial group defined in the metadata.
  \item Creates a Venn diagram to identify shared and unique OTUs among groups.
  \item For each Venn segment, generates bar plots, performs statistical analyses, and creates boxplots to visualize group-specific differences.
  \item Outputs the processed data and visualizations for further analysis.
}

Statistical tests are performed to identify significant differences in OTU abundances between groups for each Venn segment. The results are returned as data frames and visualized using boxplots.
}
\examples{
result = VenSuper.micro(ps = ps.16s,group =  "Group",num = 6)
 p7_1 <- result[[1]]
 p7_1+ scale_fill_manual(values = colset1)+ scale_color_manual(values = colset1,guide = F)
 p7_2 <- result[[3]]
 p7_2
 p8 <- result[[2]]
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Peng-Hao Xie \email{2019103106@njau.edu.cn}
}
