% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost.omics.R
\name{xgboost.omics}
\alias{xgboost.omics}
\title{XGBoost model screening of features in multi-omics data}
\usage{
xgboost.omics(ps = ps, seed = 200, top = 20, k = 5, adjust = F)
}
\arguments{
\item{ps}{A phyloseq format file used as an alternative for the input containing otu, tax, and map.}

\item{seed}{The random seed for reproducibility.}

\item{top}{The top microorganisms to consider.}

\item{k}{The number of folds for cross-validation.}

\item{adjust}{A logical value for whether to adjust XGBoost model parameters,default F.}
}
\value{
A list object including the following components:
\item{AUC}{The average accuracy of the XGBoost model.}
\item{Importance}{A data frame showing the feature importance ranked in descending order.}
}
\description{
XGBoost, one of the machine learning methods, was used to screen for characteristic
microorganisms, and the model was evaluated using k-fold cross-validation.
}
\examples{
library(dplyr)
library(ggClusterNet)
library(caret)
library(xgboost)
library(Ckmeans.1d.dp)
library(tidyverse)
library(mia)
res = xgboost.omics(ps =ps.16s, top = 50 )
accuracy = res[[1]]
accuracy
importance = res[[2]]
importance
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
