% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone.fill.ms.R
\name{zone.fill.ms}
\alias{zone.fill.ms}
\title{Frequency filtering of metabolites in repeated samples}
\usage{
zone.fill.ms(ps, method = "repeat", n = 0.667)
}
\arguments{
\item{ps}{A phyloseq format file used as an alternative for the input containing metabolite composition table,
metabolite classification table, and sample metadata.}

\item{method}{The method to use for filling zones. Default is "repeat".}

\item{n}{Metabolite appearance frequency threshold. Default is 0.667(2/3).}
}
\value{
The filled phyloseq format file after filtered based on the selected
metabolite appearance frequency threshold.
}
\description{
This function mainly filters the frequency of metabolite occurrence in duplicate samples to improve data reliability.
By default, metabolites that appear in 2/3 of duplicate samples are retained,
otherwise their abundance in each repeated sample is filled to 0.
}
\examples{
ps.ms2 = zone.fill.ms(ps = ps.ms,method = "repeat")
}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
