% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluMicro.bar.metm.R
\name{cluMicro.bar.metm}
\alias{cluMicro.bar.metm}
\title{Clustered Bar Plots with Hierarchical Clustering for Microbial Data}
\usage{
cluMicro.bar.metm(
  dist = "bray",
  otu = NULL,
  tax = NULL,
  map = NULL,
  tree = NULL,
  j = "Phylum",
  ps = NULL,
  rep = 6,
  Top = 10,
  tran = TRUE,
  hcluter_method = "complete",
  Group = "Group",
  cuttree = 3,
  group_colors = NULL
)
}
\arguments{
\item{dist}{Character string specifying the distance method (default: "bray").}

\item{otu, tax, map, tree}{Optional input tables. If NULL, values are extracted
from \code{ps}.}

\item{j}{Character string specifying the taxonomic rank to visualize
(e.g., "Phylum", "Genus"). Default is "Phylum".}

\item{ps}{A \code{phyloseq} object containing microbiome data.}

\item{rep}{Integer, number of replicates (unused, kept for compatibility).}

\item{Top}{Integer, number of top taxa to keep (by abundance). Others are
merged into "Other". Default = 10.}

\item{tran}{Logical, if TRUE transform abundances to relative values.
Default = TRUE.}

\item{hcluter_method}{Character string specifying hierarchical clustering
method. Default = "complete".}

\item{Group}{Character string specifying the grouping column in sample metadata.
Default = "Group".}

\item{cuttree}{Integer, number of clusters to cut the dendrogram into.
Default = 3.}
}
\description{
Perform hierarchical clustering on microbial community data at both the sample
level and the group-mean level, and visualize results by combining dendrograms
with stacked bar plots at a specified taxonomic rank.
}
\details{
Steps performed:
\itemize{
  \item Transform OTU table to relative abundances (if \code{tran = TRUE}).
  \item Perform hierarchical clustering at sample level and group-average level.
  \item Collapse taxa to the specified rank, retain top \code{Top} taxa and merge others.
  \item Create stacked bar plots aligned with dendrograms using \code{facet_plot()}.
}
}
\examples{
\dontrun{
res <- cluMicro.bar.metm(
  dist = "bray", ps = ps.16s, j = "Genus", Top = 10,
  tran = TRUE, hcluter_method = "complete", Group = "Group",
  cuttree = length(unique(phyloseq::sample_data(ps.16s)[["Group"]]))
)
res[[1]]  # sample-level dendrogram
res[[2]]  # sample-level dendrogram + stacked barplot
res[[3]]  # group-level dendrogram
res[[4]]  # group-level dendrogram + stacked barplot
head(res[[5]])
}

}
\author{
Tao Wen \email{2018203048@njau.edu.cn},
  Peng-Hao Xie \email{2019103106@njau.edu.cn}
}
