% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_amplicon_result_dir.R
\name{print.psSeqType}
\alias{print.psSeqType}
\title{Print method for \code{psSeqType} objects}
\usage{
\method{print}{psSeqType}(x, ...)
}
\arguments{
\item{x}{A \code{psSeqType} object.}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Print method for \code{psSeqType} objects
}
