% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomforest.trans.R
\name{randomforest.trans}
\alias{randomforest.trans}
\title{Random forest modeling for microbiome data}
\usage{
randomforest.trans(
  otu = NULL,
  tax = NULL,
  map = NULL,
  tree = NULL,
  ps = NULL,
  group = "Group",
  optimal = 20,
  rfcv = FALSE,
  nrfcvnum = 5,
  min = -1,
  max = 5,
  fill = "Level1",
  lab = "id"
)
}
\arguments{
\item{otu}{OTU/ASV table;}

\item{tax}{taxonomy table}

\item{map}{Sample metadata;}

\item{ps}{phyloseq object of microbiome}

\item{optimal}{important OTU number which selected}

\item{rfcv}{TURE or FELSE,whether need to do cross-validation}

\item{nrfcvnum}{Number of cross-validation}

\item{min}{Circle diagram inner diameter adjustment}

\item{max}{Circle diagram outer diameter adjustment}

\item{Group}{column name for groupID in map table.}
}
\value{
list contain ggplot object and table.
}
\description{
Random forest modeling for microbiome data
}
\examples{
# data form github
result = MicroRF(ps = ps,group  = "Group",optimal = 20,rfcv = TRUE,nrfcvnum = 5,min = -1,max = 5)
}
\references{
Jingying Zhang, Yong-Xin Liu, Na Zhang, Bin Hu, Tao Jin, Haoran Xu, Yuan Qin, Pengxu Yan, Xiaoning Zhang, Xiaoxuan Guo, Jing Hui, Shouyun Cao, Xin Wang, Chao Wang, Hui Wang, Baoyuan Qu, Guangyi Fan, Lixing Yuan, Ruben Garrido-Oter, Chengcai Chu & Yang Bai.
NRT1.1B is associated with root microbiota composition and nitrogen use in field-grown rice.
Nature Biotechnology, 2019(37), 6:676-684, DOI: \url{https://doi.org/10.1038/s41587-019-0104-4}
}
\author{
Contact: Tao Wen \email{2018203048@njau.edu.cn}, Yong-Xin Liu \email{yxliu@genetics.ac.cn}
}
