% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ven.Upset.metm.R
\name{Ven.Upset.metm}
\alias{Ven.Upset.metm}
\title{Generate Venn Diagrams and Upset Plots for Microbial Group Comparisons}
\usage{
Ven.Upset.metm(
  otu = NULL,
  tax = NULL,
  map = NULL,
  tree = NULL,
  ps = NULL,
  group = "Group",
  N = 0.5,
  size = 3
)
}
\arguments{
\item{otu}{An OTU table (optional). If `NULL`, uses the `ps` object.}

\item{tax}{A taxonomy table (optional). If `NULL`, uses the `ps` object.}

\item{map}{A sample metadata table (optional). If `NULL`, uses the `ps` object.}

\item{tree}{A phylogenetic tree (optional). If `NULL`, uses the `ps` object.}

\item{ps}{A `phyloseq` object containing microbiome data.}

\item{group}{A character string specifying the grouping variable in the sample metadata. Default is `"Group"`.}

\item{N}{A numeric threshold for presence-absence data. Default is `0.5`.}

\item{size}{Numeric. The scaling factor for the Upset plot. Default is `3`.}
}
\value{
A list containing:
\describe{
  \item{p}{A `ggVennDiagram` object representing the Venn diagram.}
  \item{p2}{A composite Upset plot.}
  \item{ven2}{A processed binary presence-absence data frame.}
  \item{main_plot}{The main bar plot of the Upset plot.}
  \item{side_plot}{The side bar plot of the Upset plot.}
}
}
\description{
This function creates Venn diagrams and Upset plots to compare microbial groups based on presence/absence thresholds.
}
\examples{
\dontrun{
library(phyloseq)
# Example with a phyloseq object
res = Ven.Upset.metm(ps =  ps.16s,group = "Group",N = 0.5,size = 3)
p10.1 = res[[1]]
p10.1
p10.2 = res[[2]]
p10.2
p10.3 = res[[4]]
p10.3
dat = res[[3]]
head(dat)
}

}
\author{
Tao  Wen \email{2018203048@njau.edu.cn},
Peng-Hao Xie \email{2019103106@njqu.edu.cn}
}
